% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apeglmContrast.R
\name{apeglmContrast}
\alias{apeglmContrast}
\title{Run a quick pairwise contrast using lfcShrink with apeglm}
\usage{
apeglmContrast(dds, contrast, ...)
}
\arguments{
\item{...}{Passes to \code{\link[DESeq2:lfcShrink]{lfcShrink()}}, with
\code{type = "apeglm"} and \code{coef} automatically defined. Optionally can pass
unshrunken \code{DESeqResults} via \code{res} argument here and this will set
\code{alpha}, \code{lfcThreshold} automatically.}
}
\description{
Log2 fold change shrinkage via \code{\link[DESeq2:lfcShrink]{lfcShrink()}}
currently only supports \code{coef} but not \code{contrast} argument, so we're defining
a wrapper function here that sets \code{coef} internally automatically.
}
\details{
Generates model matrices internally via
\code{\link[stats:model.matrix]{model.matrix()}} and then dynamically sets reference
factor levels, as recommended by DESeq2 vignette.

Runs \code{\link[DESeq2:nbinomWaldTest]{nbinomWaldTest()}} via
\code{\link[DESeq2:DESeq2]{DESeq()}}, followed by \code{\link[DESeq2:lfcShrink]{lfcShrink()}}.
}
\note{
Updated 2019-09-17.
}
\examples{
dds <- DESeq2::makeExampleDESeqDataSet(n = 1000L, m = 12L)
dds$condition <- factor(rep(LETTERS[seq_len(4L)], each = 3L))
dds <- DESeq(dds)
resultsNames(dds)

## Contrast C vs. B.
contrast <- c(factor = "condition", numerator = "C", denominator = "B")

## Unshrunken DESeqResults.
res <- results(dds, contrast = contrast)
class(res)
lfcShrinkType(res)

## Shrunken DESeqResults, using apeglm via `lfcShrink()`.
shrink <- apeglmContrast(dds = dds, contrast = contrast)
class(shrink)
lfcShrinkType(shrink)
}
\seealso{
\itemize{
\item "Extended section on shrinkage estimators" section of DESeq2 vignette,
which explains how to manually define \code{coef} argument which can
be used with apeglm \code{\link[DESeq2:lfcShrink]{lfcShrink()}}.
\item \code{\link[stats:model.matrix]{stats::model.matrix()}}.
\item \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.
\item \code{\link[apeglm:apeglm]{apeglm::apeglm()}}.
\item \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
}
}
