% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/apeglmResults-methods.R
\name{apeglmResults}
\alias{apeglmResults}
\alias{apeglmResults,DESeqDataSet-method}
\title{Run a quick pairwise contrast using lfcShrink with apeglm}
\usage{
apeglmResults(object, ...)

\S4method{apeglmResults}{DESeqDataSet}(object, contrast, res, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{contrast}{\code{character(3)}.
Pairwise contrast vector:
\enumerate{
\item \code{factor}: Grouping factor. Corresponds to column name in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}}.
\item \code{numerator}: Numerator samples.
\item \code{denominator}: Denominator samples.
}

Numerator and denominator values correspond to grouping factor column.
See \code{\link[=results]{results()}} for details. Note that we're intentionally being more
strict about the input format here.}

\item{res}{\code{DESeqResults}.
Results containing unshrunken LFC values, generated with \code{results()}.}
}
\value{
\code{DESeqResults}, with apeglm adaptive shrinkage applied to fold
change values.
}
\description{
Wrapper function that helps set up \code{\link[DESeq2:lfcShrink]{lfcShrink()}} to
shrink LFC values for a pairwise contrast via apeglm, without having to
manually relevel factor reference levels to use the required \code{coef} argument.
}
\details{
Dynamically sets reference factor levels, as recommended by DESeq2 vignette.
Matches \code{contrast} input internally to corresponding \code{coef} corresponding
to values in \code{\link[DESeq2:results]{resultsNames()}}.

Runs \code{\link[DESeq2:nbinomWaldTest]{nbinomWaldTest()}} via
\code{\link[DESeq2:DESeq]{DESeq()}}, followed by \code{\link[DESeq2:lfcShrink]{lfcShrink()}}.
}
\note{
Updated 2020-08-17.
}
\examples{
## DESeqDataSet ====
if (requireNamespace("apeglm", quietly = TRUE)) {
    dds <- DESeq2::makeExampleDESeqDataSet(n = 1000L, m = 12L)
    dds$condition <- factor(rep(LETTERS[seq_len(4L)], each = 3L))
    dds <- DESeq2::DESeq(dds)
    resultsNames(dds)

    ## Contrast C vs. B.
    contrast <- c(factor = "condition", numerator = "C", denominator = "B")

    ## Unshrunken DESeqResults.
    res <- DESeq2::results(dds, contrast = contrast)
    class(res)
    lfcShrinkType(res)

    ## Shrunken DESeqResults, using apeglm via `lfcShrink()`.
    shrink <- apeglmResults(
        object = dds,
        contrast = contrast,
        res = res
    )
    class(shrink)
    lfcShrinkType(shrink)
}
}
\seealso{
\itemize{
\item "Extended section on shrinkage estimators" section of DESeq2 vignette,
which explains how to manually define \code{coef} argument which can
be used with apeglm \code{\link[DESeq2:lfcShrink]{lfcShrink()}}.
\item \code{\link[apeglm:apeglm]{apeglm::apeglm()}}.
\item \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.
\item \code{\link[DESeq2:results]{DESeq2::resultsNames()}}.
\item \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\item \code{stats::model.matrix()}.
}
}
