% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{as}
\alias{as}
\alias{coerce}
\alias{coerce,DESeqAnalysis,DESeqDataSet-method}
\alias{coerce,DESeqAnalysis,DESeqTransform-method}
\alias{coerce,DESeqAnalysis,DESeqResultsList-method}
\title{Force an object to belong to a class}
\description{
Force an object to belong to a class
}
\note{
Updated 2021-03-09.
}
\section{DESeqAnalysis}{


Supported coercion methods will extract any of these internal objects:
\itemize{
\item \code{DESeqDataSet}.
\item \code{DESeqTransform}.
\item \code{DESeqResultsList}.
}
}

\examples{
data(deseq)

## DESeqAnalysis ====
dds <- as(deseq, "DESeqDataSet")
print(dds)
dt <- as(deseq, "DESeqTransform")
print(dt)
resList <- as(deseq, "DESeqResultsList")
print(names(resList))
}
\author{
Michael Steinbaugh
}
