% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMeanThreshold-methods.R
\name{baseMeanThreshold}
\alias{baseMeanThreshold}
\alias{baseMeanThreshold<-}
\alias{baseMeanThreshold,Annotated-method}
\alias{baseMeanThreshold<-,Annotated,numeric-method}
\title{Base mean threshold}
\usage{
baseMeanThreshold(object, ...)

baseMeanThreshold(object, ...) <- value

\S4method{baseMeanThreshold}{Annotated}(object)

\S4method{baseMeanThreshold}{Annotated,numeric}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)}.
Base mean cutoff level.
}
\description{
Base mean threshold
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
object <- deseq
baseMeanThreshold(object) <- 100L
baseMeanThreshold(object)
}
