% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce-methods.R
\name{as.DESeqDataSet}
\alias{as.DESeqDataSet}
\alias{as.DESeqTransform}
\alias{coerce}
\alias{as.DESeqDataSet,DESeqAnalysis-method}
\alias{as.DESeqTransform,DESeqAnalysis-method}
\alias{coerce,DESeqAnalysis,DESeqDataSet-method}
\alias{coerce,DESeqAnalysis,DESeqTransform-method}
\alias{coerce,DESeqAnalysis,DESeqResultsList-method}
\title{Coercion methods}
\usage{
as.DESeqDataSet(x, ...)

as.DESeqTransform(x, ...)

\S4method{as.DESeqDataSet}{DESeqAnalysis}(x)

\S4method{as.DESeqTransform}{DESeqAnalysis}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, of desired conversion class.
}
\description{
Coercion methods
}
\note{
Updated 2022-03-30.
}
\section{DESeqAnalysis}{


Supported coercion methods will extract any of these internal objects:
\itemize{
\item \code{DESeqDataSet}.
\item \code{DESeqTransform}.
\item \code{DESeqResultsList}.
}
}

\examples{
data(deseq)

## DESeqAnalysis ====
dds <- as.DESeqDataSet(deseq)
class(dds)
dt <- as.DESeqTransform(deseq)
class(dt)
}
