% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\name{combine}
\alias{combine}
\alias{combine,DESeqAnalysis,DESeqAnalysis-method}
\title{Combining or merging different Bioconductor data structures}
\usage{
combine(x, y, ...)

\S4method{combine}{DESeqAnalysis,DESeqAnalysis}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
A single value of the same class as the most specific common ancestor
  (in class terms) of the input values. This will contain the
  appropriate combination of the data in the input values.
}
\description{
Combining or merging different Bioconductor data structures
}
\details{
Combines the results from 2 separate \code{DESeqAnalysis} objects. Note that the
internal \code{DESeqDataSet} and \code{DESeqTransform} objects slotted in \code{x} and \code{y}
must be identical. The respective results must not contain any intersecting
names. The \code{lfcShrink} slot is optional but must be either defined in both
objects or \code{NULL} in both objects.
}
\note{
Updated 2019-11-19.
}
\examples{
data(deseq)

x <- deseq
y <- deseq
resultsNames(x) <- paste0("x_", resultsNames(x))
resultsNames(y) <- paste0("y_", resultsNames(y))

object <- combine(x, y)
resultsNames(object)
}
