% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastName-methods.R
\name{contrastName}
\alias{contrastName}
\alias{contrastName<-}
\alias{contrastName,DESeqResults-method}
\alias{contrastName<-,DESeqResults,character-method}
\alias{contrastName,DESeqAnalysis-method}
\title{Contrast name}
\usage{
contrastName(object, ...)

contrastName(object, ...) <- value

\S4method{contrastName}{DESeqResults}(object, .format = c("resultsNames", "title"), .useStash = TRUE)

\S4method{contrastName}{DESeqResults,character}(object) <- value

\S4method{contrastName}{DESeqAnalysis}(object, i, ...)
}
\arguments{
\item{object}{Object.}

\item{.format}{\code{character(1)}.
Name format to return:
\itemize{
\item \code{resultsNames}: Attempt to matching the conventions in
\code{\link[DESeq2:results]{resultsNames()}}.
\item \code{title}: Human readable, for plot titles and/or table captions.
}}

\item{.useStash}{\code{logical(1)}.
Check for \code{contrastName} metadata stash in \code{DESeqResults} object. Intended
for use with \code{DESeqAnalysis} methods.}

\item{value}{Value to assign.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Contrast name.
}
\description{
Contrast name
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
contrastName(deseq, i = 1L)
}
\seealso{
\itemize{
\item \code{\link[=contrastNames]{contrastNames()}}.
\item \code{\link[DESeq2:results]{resultsNames()}}.
}
}
