% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastSamples-methods.R
\docType{methods}
\name{contrastSamples}
\alias{contrastSamples}
\alias{contrastSamples,DESeqAnalysis-method}
\title{Match the Samples Used to Define Contrast}
\usage{
contrastSamples(object, ...)

\S4method{contrastSamples}{DESeqAnalysis}(object, results = 1L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
\code{character}. Sample identifiers (corresponding to columns).
}
\description{
Match the samples in a \code{DESeqDataSet} used to define contrast in a corresponding
\code{DESeqResults} object. Note that this only works for simple (e.g. pairwise)
contrasts and will intentionally error for more complex comparisons.
}
\note{
\strong{EXPERIMENTAL METHOD}. This approach will error intentionally for
\code{DESeqResults} objects generated with complex contrasts (e.g. interaction
effect or LRT). It may be removed in a future release if this approach
is too error-prone.
}
\examples{
data(deseq)

## DESeqAnalysis ====
contrastSamples(deseq, results = 1L)
}
