% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-methods.R
\docType{methods}
\name{correlation}
\alias{correlation}
\alias{correlation,DESeqResults,DESeqResults-method}
\title{Correlation}
\usage{
correlation(x, y, ...)

\S4method{correlation}{DESeqResults,DESeqResults}(x, y,
  col = "log2FoldChange", method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{col}{\code{character(1)}.
Column name.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)} or \code{matrix}.
}
\description{
Correlation
}
\note{
Updated 2019-11-08.
}
\examples{
data(deseq, package = "acidtest")

## DESeqResults ====


## DESeqAnalysis ====

}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{stats::cor.test()}.
\item \code{stats::sd()}.
\item \code{S4Vectors::cor()}.
\item https://stats.stackexchange.com/questions/24980
}
}
