% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dds_list}
\alias{dds_list}
\title{List of DESeqDataSet example objects}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(dds_list)
}
\description{
List of DESeqDataSet example objects
}
\note{
Updated 2021-06-29.
}
\examples{
data(dds_list)
dds_list
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
