% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg-methods.R
\docType{methods}
\name{deg}
\alias{deg}
\alias{deg,DESeqResults-method}
\alias{deg,DESeqAnalysis-method}
\title{Differentially Expressed Genes}
\usage{
\S4method{deg}{DESeqResults}(object, alpha = NULL, lfcThreshold = NULL,
  direction = c("both", "up", "down"))

\S4method{deg}{DESeqAnalysis}(object, results = 1L,
  direction = c("both", "up", "down"))
}
\arguments{
\item{object}{Object.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
\code{character}. Gene identifiers.
}
\description{
Differentially Expressed Genes
}
\examples{
data(deseq)

## DESeqAnalysis ====
deg(deseq)
}
