% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degIntersection-methods.R
\name{degIntersection}
\alias{degIntersection}
\alias{degIntersection,DESeqAnalysis-method}
\title{Intersection of differentially expressed genes}
\usage{
degIntersection(object, ...)

\S4method{degIntersection}{DESeqAnalysis}(
  object,
  i = NULL,
  direction = c("up", "down"),
  return = c("matrix", "count", "ratio")
)
}
\arguments{
\item{object}{Object.}

\item{i}{\code{character}, \code{numeric}, or \code{NULL}.
Names or range of results.
If set \code{NULL}, include all results.}

\item{direction}{\code{character(1)}.
Include "up" or "down" directions.
Must be directional, and intentionally does not support "both".}

\item{return}{\code{character(1)}.
Return intersection matrix, count per contrast, or ratio.}

\item{...}{Passthrough arguments to \code{\link[=deg]{deg()}}.}
}
\value{
\code{integer}.
Integer denoting the number of intersections, sorted from highest to
lowest. Gene identifiers are defined as the names.
}
\description{
Intersection of differentially expressed genes
}
\note{
Updated 2020-08-18.
}
\examples{
data(deseq)

## DESeqAnalysis ====
mat <- degIntersection(deseq, return = "matrix")
head(mat)

count <- degIntersection(deseq, return = "count")
head(count)

ratio <- degIntersection(deseq, return = "ratio")
head(ratio)
}
