% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deseq}
\alias{deseq}
\title{DESeq2 example analysis}
\format{
An object of class \code{DESeqAnalysis} with 500 rows and 12 columns.
}
\usage{
data(deseq)
}
\description{
Minimal DESeq2 differential expression analysis.
}
\note{
Updated 2021-06-29.
}
\examples{
data(deseq)
deseq
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
