% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqAnalysis-class}
\alias{DESeqAnalysis-class}
\title{DESeq2 differential expression analysis}
\value{
\code{DESeqAnalysis}.
Contains a \code{DESeqDataSet}, \code{DESeqTransform}, and corresponding
\code{DESeqResults} list.
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2. This class is essentially a \code{list} with validity checks
to ensure \code{DESeqTransform} and \code{DESeqResults} correspond to the
\code{DESeqDataSet}.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{DESeqDataSet}.}

\item{\code{transform}}{\code{DESeqTransform}.}

\item{\code{results}}{\code{list}.
One or more unshrunken \code{DESeqResults}.}

\item{\code{lfcShrink}}{\code{list}.
\emph{Optional}. One or more shrunken \code{DESeqResults}. If set, must correspond to
those defined in \code{results}.}
}}

\section{DESeqDataSet}{


We recommend generating the \code{DESeqDataSet} by coercion from \code{bcbioRNASeq}
object using \code{as(dds, "bcbioRNASeq")}. Don't use the \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}
or \code{DESeq2::DESeqDataSetFromMatrix()} constructors to generate the
\code{DESeqDataSet} object.
}

\section{DESeqTransform}{


Object containing variance-stabilized counts. We recommend slotting the
return from either \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}} or
\code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
}

\section{DESeqResults}{


Don't modify any of the \code{DESeqResults} objects manually. This includes
rearranging the rows or dropping genes without adjusted P values. We'll take
care of this automatically in supported methods.
}

\seealso{
\code{\link[=DESeqAnalysis]{DESeqAnalysis()}}.
}
\author{
Michael Steinbaugh
}
