% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{DESeqAnalysis-package}
\title{DESeqAnalysis}
\description{
Toolkit for performing differential expression with
\href{http://bioconductor.org/packages/DESeq2/}{DESeq2}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://deseqanalysis.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/DESeqAnalysis/}
  \item Report bugs at \url{https://github.com/acidgenomics/DESeqAnalysis/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
