% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{DESeqAnalysisList}
\alias{DESeqAnalysisList}
\title{List containing related DESeq2 analyses}
\usage{
DESeqAnalysisList(...)
}
\arguments{
\item{...}{\code{DESeqAnalysis} objects.}
}
\value{
\code{DESeqAnalysisList}.
}
\description{
List containing related DESeq2 analyses
}
\note{
Updated 2019-08-20.
}
\examples{
data(deseq)
x <- DESeqAnalysisList(deseq)
x
}
