% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deseqMinimal}
\alias{deseqMinimal}
\title{Minimal DESeq2 example analysis}
\format{
An object of class \code{DESeqAnalysis} with 1000 rows and 12 columns.
}
\usage{
data(deseqMinimal)
}
\description{
Minimal DESeq2 example analysis
}
\note{
Updated 2022-05-17.
}
\examples{
data(deseqMinimal)
print(deseqMinimal)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
