% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllGenerators.R
\name{DESeqResultsList}
\alias{DESeqResultsList}
\alias{DESeqResultsList,DESeqAnalysis-method}
\alias{DESeqResultsList,DESeqAnalysisList-method}
\alias{DESeqResultsList,SimpleList-method}
\alias{DESeqResultsList,list-method}
\alias{DESeqResultsList,missing-method}
\title{List containing related DESeqResults objects}
\usage{
DESeqResultsList(object, ...)

\S4method{DESeqResultsList}{DESeqAnalysis}(object, lfcShrink = NULL, quiet = FALSE)

\S4method{DESeqResultsList}{DESeqAnalysisList}(object, ...)

\S4method{DESeqResultsList}{SimpleList}(object)

\S4method{DESeqResultsList}{list}(object)

\S4method{DESeqResultsList}{missing}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{lfcShrink}{\code{logical(1)} or \code{NULL}.
Use shrunken log2 fold change (LFC) values.
If \code{NULL}, inherits value defined in \code{lfcShrink()}.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}
}
\value{
\code{DESeqResultsList}.
}
\description{
List containing related DESeqResults objects
}
\note{
Updated 2021-03-15.
}
\examples{
data(deseq)
x <- DESeqResultsList(deseq)
print(names(x))
}
