% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqResultsTables-class}
\alias{DESeqResultsTables-class}
\title{DESeq2 Differential Expression Results Tables}
\value{
\code{DESeqResultsTables}, containing the original, unmodified
\code{DESeqResults} \code{DataFrame} along with the corresponding differentially
expressed genes and gene-level metadata (rowRanges).
}
\description{
\code{DESeqResults} object with \code{DataFrame} subsets and file path information.
}
\section{Slots}{

\describe{
\item{\code{results}}{\code{DESeqResults}. Original unmodified \code{DESeqResults}. Should
contain all genes, including those with \code{NA} adjusted \emph{P} values.}

\item{\code{deg}}{\code{list}. Differentially expressed genes. Contains \code{character}
vectors of genes that are upregulated (\code{up}) or downregulated (\code{down}).
Values map to the \code{rownames} of the internal \code{DESeqResults}. These are
genes that pass \code{alpha} and \code{lfcThreshold} cutoffs set in
\code{DESeq2::results()} call.}

\item{\code{counts}}{\code{matrix}. Size-factor adjusted (normalized) counts matrix from
the \code{DESeqDataSet}. This matrix is returned by \code{DESeq2::counts(object, normalized = TRUE)}. These should NOT be the log2 variance-stabilized
counts defined in the \code{DESeqTransform} object.}

\item{\code{rowRanges}}{\code{GRanges}. Row annotations.}

\item{\code{sampleNames}}{\code{character}. Human-friendly sample names. Must contain
\code{names()} that map to the \code{colnames()} of the \code{DESeqDataSet}.}

\item{\code{metadata}}{\code{list}. Metadata. Contains file paths and information on
whether we're writing locally or to Dropbox.}
}}

\section{Obtaining results from DESeq2}{


It is recommended to specify the \code{contrast} argument as \code{character}:
\enumerate{
\item Design matrix factor of interest.
\item Numerator for LFC (expt).
\item Denominator for LFC (control).
}

For example, to get the relative change in gene expression between mutant
and wildtype samples, here's how to set the contrast vector:\preformatted{c(
    factor = "genotype",
    numerator = "mutant",
    denominator = "wildtype"
)
}
}

\section{Log fold change threshold cutoffs}{


Refer to \code{DESeq2::results()} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based on
expected fold changes. In addition, the "Hypothesis tests with thresholds on
effect size" section in the DESeq2 paper provides additional explanation.

Don't apply \emph{post hoc} LFC threshold filtering to obtain results tables, as
this will destroy the meaning of the adjusted \emph{P} values computed. If you are
expecting a biological effect size past a particular threshold, set
\code{lfcThreshold}, but be conservative.

This \href{https://support.bioconductor.org/p/101504/}{thread} on the Bioconductor forums explains how \code{DESeq2::results()}
should be called with regard to LFC cutoffs in nice detail.
}

\seealso{
\itemize{
\item \code{DESeq2::results()}.
\item \code{markdown()}, \code{write()}.
}

\code{DESeqResultsTables()}.

Other S4 classes: \code{\link{DESeqAnalysis-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 classes}
