% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,DESeqAnalysis,character,missingOrNULL-method}
\alias{export,DESeqAnalysis,missingOrNULL,missingOrNULL-method}
\alias{export,DESeqDataSet,character,missingOrNULL-method}
\alias{export,DESeqDataSet,missingOrNULL,missingOrNULL-method}
\title{Export}
\usage{
export(object, con, format, ...)

\S4method{export}{DESeqAnalysis,character,missingOrNULL}(
  object,
  con,
  format,
  compress = getOption(x = "acid.export.compress", default = FALSE),
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{DESeqAnalysis,missingOrNULL,missingOrNULL}(object, con, format, name = NULL, dir, ...)

\S4method{export}{DESeqDataSet,character,missingOrNULL}(
  object,
  con,
  format,
  compress = getOption(x = "acid.export.compress", default = FALSE),
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{DESeqDataSet,missingOrNULL,missingOrNULL}(object, con, format, name = NULL, dir, ...)
}
\arguments{
\item{object}{Object.}

\item{con}{\code{character(1)}.
Directory path.}

\item{format}{\code{character(1)}, \code{missing}, or \code{NULL}.
Output file format extension.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export
}
\details{
Size-factor normalized coutns and FPKM values are calculated on the fly and
exported automatically.
}
\note{
Updated 2022-05-17.
}
\examples{
data(deseq)

## DESeqAnalysis ====
con <- AcidBase::tempdir2()
x <- export(deseq, con = con)
print(x)
AcidBase::unlink2(con)
}
