% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{extract}
\alias{extract}
\alias{[,DESeqAnalysis,ANY,ANY,ANY-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{[}{DESeqAnalysis,ANY,ANY,ANY}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{drop}{relevant for matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }
}
\value{
\code{DESeqAnalysis}.
}
\description{
Extract genes by row and samples by column.
}
\note{
Updated 2020-10-28.
}
\examples{
data(deseq)
object <- deseq
dim(object)

genes <- head(rownames(object), 50L)
head(genes)
samples <- head(colnames(object), 2L)
head(samples)

x <- object[genes, samples]
print(x)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.
}
