% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPairwiseContrasts.R
\name{importPairwiseContrasts}
\alias{importPairwiseContrasts}
\title{Import pairwise contrasts from user-defined file}
\usage{
importPairwiseContrasts(file, groupCol = "group",
  numeratorCol = "numerator", denominatorCol = "denominator",
  namesCol = "description")
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{groupCol}{\code{character(1)}.
Group column name in \code{\link[SummarizedExperiment:colData]{colData()}} of
\code{DESeqDataSet}. Corresponds to value in \code{\link[DESeq2:design]{design()}}.}

\item{numeratorCol}{\code{character(1)}.
Numerator column name in contrasts file.}

\item{denominatorCol}{\code{character(1)}.
Denominator column name in contrasts file.}

\item{namesCol}{\code{character(1)}.
Column in contrasts file to used to define the names of the contrast list.
Values will be sanitized into \code{\link[basejump:snakeCase]{snakeCase()}}.}
}
\value{
\code{list}.
Named list of pairwise character vectors containing: "group", "numerator",
"denominator".
}
\description{
Import pairwise contrasts from user-defined file
}
\examples{
file <- system.file("extdata/contrasts.csv", package = "DESeqAnalysis")
x <- importPairwiseContrasts(file)
print(x)
}
\seealso{
\code{DESeq2::results()}.
}
