% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestingGroups-methods.R
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{interestingGroups,DESeqAnalysis-method}
\alias{interestingGroups<-,DESeqAnalysis,character-method}
\alias{interestingGroups<-,DESeqAnalysis,NULL-method}
\title{Interesting groups}
\usage{
interestingGroups(object, ...)

interestingGroups(object, ...)  <- value

\S4method{interestingGroups}{DESeqAnalysis}(object)

\S4method{interestingGroups}{DESeqAnalysis,character}(object) <- value

\S4method{interestingGroups}{DESeqAnalysis,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Interesting groups
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
intgroup <- interestingGroups(deseq)
print(intgroup)
## Assignment support.
interestingGroups(deseq) <- intgroup
interestingGroups(deseq)
}
