% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lfcShrinkType-methods.R
\name{lfcShrinkType}
\alias{lfcShrinkType}
\alias{lfcShrinkType,DESeqResults-method}
\alias{lfcShrinkType,DESeqAnalysis-method}
\title{Shrunken log2 fold change (LFC) type}
\usage{
lfcShrinkType(object, ...)

\S4method{lfcShrinkType}{DESeqResults}(object)

\S4method{lfcShrinkType}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)} or \code{NULL}.
Name of adaptive shrinkage method applied (e.g. "apeglm" or "ashr").
}
\description{
Shrunken log2 fold change (LFC) type
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
lfcShrinkType(deseq)
}
\seealso{
\code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.
}
