% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcThreshold-methods.R
\name{lfcThreshold}
\alias{lfcThreshold}
\alias{lfcThreshold<-}
\alias{lfcThreshold,DESeqResults-method}
\alias{lfcThreshold,DESeqAnalysis-method}
\alias{lfcThreshold<-,DESeqAnalysis,numeric-method}
\title{Log2 fold change threshold}
\usage{
lfcThreshold(object, ...)

lfcThreshold(object, ...) <- value

\S4method{lfcThreshold}{DESeqResults}(object)

\S4method{lfcThreshold}{DESeqAnalysis}(object)

\S4method{lfcThreshold}{DESeqAnalysis,numeric}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)}.
}
\description{
Log2 fold change threshold
}
\note{
Updated 2019-08-06.
}
\examples{
data(deseq)

## DESeqAnalysis ====
lfcThreshold(deseq) <- 0.2
lfcThreshold(deseq)
}
