% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markdownTables-methods.R
\name{markdownTables}
\alias{markdownTables}
\alias{markdownTables,DFrameList-method}
\alias{markdownTables,DESeqAnalysis-method}
\alias{markdownTables,DESeqResults-method}
\title{Markdown tables}
\usage{
markdownTables(object, ...)

\S4method{markdownTables}{DFrameList}(object, contrastName = NULL, n = 10L)

\S4method{markdownTables}{DESeqAnalysis}(object, i, n = 10L)

\S4method{markdownTables}{DESeqResults}(object, contrastName = NULL, n = 10L)
}
\arguments{
\item{object}{Object.}

\item{contrastName}{\code{character(1)} or \code{NULL}.
Contrast name.}

\item{n}{\code{integer(1)}.
Number of genes (per direction) to report.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{...}{Additional arguments.}
}
\value{
\code{kable}.
Markdown tables.
}
\description{
Markdown tables
}
\note{
Updated 2022-05-24.
}
\examples{
data(deseq)

## DESeqAnalysis ====
markdownTables(deseq, i = 1L, n = 5L)

## DESeqResults 'resultsTables()' return ====
res <- results(deseq, i = 1L)
resTbl <- resultsTables(res)
markdownTables(resTbl, n = 5L)
}
