% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchFilesToMetadata.R
\name{matchFilesToMetadata}
\alias{matchFilesToMetadata}
\title{Match user metadata to file names used for tximport}
\usage{
matchFilesToMetadata(files, metadata)
}
\arguments{
\item{files}{\code{character}.
Quant file paths passed to \code{\link[tximport:tximport]{tximport()}}.}

\item{metadata}{\code{data.frame}.
User-defined metadata. The function assumes that sample identifiers are
defined in the first metadata column.}
}
\value{
\code{data.frame}.
Contains 1:1 file to metadata sample identifier mappings.
}
\description{
Match user metadata to file names used for tximport
}
\note{
Updated 2019-10-09.
}
\examples{
files <- file.path(
    "salmon",
    paste(seq_len(4L), "sample", LETTERS[seq_len(4L)], sep = "-"),
    "quant.sf"
)
metadata <- data.frame(
    sampleID = paste(seq_len(4L), "sample", LETTERS[seq_len(4L)], sep = "_"),
    condition = rep(LETTERS[seq_len(2L)], times = 2L)
)
## The function will match regardless of row order in user metadata.
metadata <- metadata[nrow(metadata):1L, ]
print(files)
print(metadata)
matchFilesToMetadata(files = files, metadata = metadata)
}
