% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMetadataToFiles.R
\name{matchMetadataToFiles}
\alias{matchMetadataToFiles}
\title{Match user metadata to file names used for tximport}
\usage{
matchMetadataToFiles(metadata, files)
}
\arguments{
\item{metadata}{\code{data.frame}.
User-defined metadata. The function assumes that sample identifiers are
defined in the first metadata column.}

\item{files}{\code{character}.
Quant file paths passed to \code{\link[tximport:tximport]{tximport()}}.
Sanitize return from \code{\link[=prepareTximportFiles]{prepareTximportFiles()}} is recommended.}
}
\value{
\code{data.frame}.
Modified metadata frame, with updated sample identifiers in slotted in
first column. Original values are stashed in "originalSampleID".
}
\description{
Match user metadata to file names used for tximport
}
\note{
Updated 2020-08-04.
}
\examples{
metadata <- data.frame(
    sampleID = paste(seq_len(4L), "sample", LETTERS[seq_len(4L)], sep = "_"),
    condition = rep(LETTERS[seq_len(2L)], times = 2L)
)
files <- file.path(
    "salmon",
    paste(seq_len(4L), "sample", LETTERS[seq_len(4L)], sep = "-"),
    "quant.sf"
)
## The function will match regardless of row order in user metadata.
files <- rev(files)
print(metadata)
print(files)
matchMetadataToFiles(metadata = metadata, files = files)
}
