% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotBaseMean-methods.R
\name{plotBaseMean}
\alias{plotBaseMean}
\alias{plotBaseMean,DESeqAnalysis-method}
\alias{plotBaseMean,DESeqDataSet-method}
\alias{plotBaseMean,DESeqResults-method}
\title{Plot base mean distribution}
\usage{
plotBaseMean(object, ...)

\S4method{plotBaseMean}{DESeqAnalysis}(object, ...)

\S4method{plotBaseMean}{DESeqDataSet}(object, ...)

\S4method{plotBaseMean}{DESeqResults}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{nonzero}{\code{logical(1)}.
Remove zero-count features (genes).}

\item{summary}{\code{logical(1)}.
Include distribution summary statistics as lines on the plot.}
}
\value{
\code{ggplot}.
}
\description{
The base mean is the mean of normalized counts of all samples, normalizing
for sequencing depth.
}
\section{Functions}{
\itemize{
\item \code{plotBaseMean,DESeqAnalysis-method}: Passes to \code{DESeqDataSet} method.

\item \code{plotBaseMean,DESeqDataSet-method}: Generates row means of normalized counts.
This value corresponds to the \code{baseMean} column of \code{DESeqResults}.
Passes to \code{numeric} method.

\item \code{plotBaseMean,DESeqResults-method}: Uses \code{baseMean} column of results.
Passes to \code{numeric} method.
}}

\note{
Updated 2022-05-17.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotBaseMean(deseq)
}
\seealso{
\itemize{
\item https://support.bioconductor.org/p/75244/
\item \code{summary()}.
}
}
