% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGHeatmap-methods.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqAnalysis-method}
\title{Differentially expressed gene heatmap}
\usage{
\S4method{plotDEGHeatmap}{DESeqAnalysis}(object, results,
  contrastSamples = FALSE, direction = c("both", "up", "down"),
  scale = c("row", "column", "none"), clusteringMethod = "ward.D2",
  clusterRows = TRUE, clusterCols = TRUE, interestingGroups = NULL,
  showRownames = FALSE, showColnames = TRUE, treeheightRow = 50L,
  treeheightCol = 50L, color = viridis::viridis,
  legendColor = viridis::viridis, borderColor = NULL, title = NULL,
  ...)
}
\arguments{
\item{object}{Object.}

\item{results}{\code{character(1)} or \code{integer(1)}. Name or positive of
\code{DESeqResults}.}

\item{contrastSamples}{\code{logical(1)}. \strong{Experimental}. Only include the samples
used to define the contrast passed to \code{DESeq2::results}. This setting
will break for complex DESeq2 contrasts (e.g. interaction effect).}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\description{
This function is an extension of \code{plotHeatmap} that is optimized
for automatic handling differentially expressed genes, rather than requiring
manual input of a gene vector. All of the optional parameters for
\code{plotHeatmap} are also supported by this function.
}
\details{
To adjust the annotation columns, modify the
\link[SummarizedExperiment:colData]{colData} of the \code{counts} argument, which must
contain a \code{SummarizedExperiment}.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGHeatmap(deseq, results = 1L)
}
