% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGStackedBar-methods.R
\name{plotDEGStackedBar}
\alias{plotDEGStackedBar}
\alias{plotDEGStackedBar,DESeqAnalysis-method}
\title{Stacked bar plot of DEGs}
\usage{
plotDEGStackedBar(object, ...)

\S4method{plotDEGStackedBar}{DESeqAnalysis}(
  object,
  i = NULL,
  direction = c("both", "up", "down"),
  fill = getOption(x = "acid.fill.discrete", default = getOption(x =
    "ggplot2.discrete.fill", default = acidplots::scale_fill_synesthesia_d())),
  flip = TRUE
)
}
\arguments{
\item{object}{Object.}

\item{i}{\code{character}, \code{numeric}, or \code{NULL}.
Names or range of results.
If set \code{NULL}, include all results.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Stacked bar plot of DEGs
}
\note{
Updated 2020-08-12.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGStackedBar(deseq)
}
