% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDegUpset-methods.R
\name{plotDegUpset}
\alias{plotDegUpset}
\alias{plotDegUpset,DESeqAnalysis-method}
\title{UpSet plot of directional DEG intersections across contrasts}
\usage{
plotDegUpset(object, ...)

\S4method{plotDegUpset}{DESeqAnalysis}(object, i = NULL, direction = c("both", "up", "down"), ...)
}
\arguments{
\item{object}{Object.}

\item{i}{\code{character}, \code{numeric}, or \code{NULL}.
Names or range of results.
If set \code{NULL}, include all results.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Passthrough arguments to \code{AcidPlots::plotUpset()}.}
}
\value{
Plot.
}
\description{
UpSet plot of directional DEG intersections across contrasts
}
\note{
Updated 2022-05-18.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDegUpset(deseq)
}
