% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,DESeqAnalysis-method}
\title{Plot Gene Expression}
\usage{
\S4method{plotGene}{DESeqAnalysis}(object, genes,
  interestingGroups = NULL, medianLine = TRUE,
  color = getOption("basejump.discrete.color", NULL),
  legend = getOption("basejump.legend", TRUE), style = c("facet",
  "wide"))
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}. Gene identifiers. It is considered better practice
to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003")
and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{style}{\code{string}. Plot style.}
}
\value{
\code{ggplot}.

\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot Gene Expression
}
\examples{
data(deseq)
object <- deseq

g2s <- Gene2Symbol(as(object, "DESeqDataSet"))
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

plotGene(object, genes = geneIDs, style = "facet")
plotGene(object, genes = geneNames, style = "wide")
}
