% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotHeatmap-methods.R
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotHeatmap,DESeqAnalysis-method}
\alias{plotCorrelationHeatmap,DESeqAnalysis-method}
\alias{plotQuantileHeatmap,DESeqAnalysis-method}
\title{Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

plotHeatmap(object, ...)

plotQuantileHeatmap(object, ...)

\S4method{plotHeatmap}{DESeqAnalysis}(object, ...)

\S4method{plotCorrelationHeatmap}{DESeqAnalysis}(object, ...)

\S4method{plotQuantileHeatmap}{DESeqAnalysis}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
Heatmap
}
\note{
Updated 2021-03-15.
}
\section{Scaling}{



Here we're scaling simply by calculating the standard score (z-score).
\itemize{
\item mu: mean.
\item sigma: standard deviation.
\item x: raw score (e.g. count matrix).
\item z: standard score (z-score).
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{z = (x - mu) / sigma
}\if{html}{\out{</div>}}

See also:
\itemize{
\item \code{pheatmap:::scale_rows()}.
\item \code{scale()} for additional scaling approaches.
}

}

\section{Hierarchical clustering}{



Row- and column-wise hierarchical clustering is performed when \code{clusterRows}
and/or \code{clusterCols} are set to \code{TRUE}. Internally, this calls \code{hclust()},
and defaults to the Ward method.

Automatic hierarchical clustering of rows and/or columns can error for some
datasets. When this occurs, you'll likely see this error:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Error in hclust(d, method = method) :
NA/NaN/Inf in foreign function call
}\if{html}{\out{</div>}}

In this case, either set \code{clusterRows} and/or \code{clusterCols} to \code{FALSE}, or
you can attempt to pass an \code{hclust} object to these arguments. This is
recommended as an alternate approach to be used with \code{pheatmap()}, which is
called internally by our plotting code. Here's how this can be accomplished:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mat <- assay(mat)
dist <- dist(mat)
hclust <- hclust(dist, method = "ward.D2")
}\if{html}{\out{</div>}}

}

\examples{
data(deseq)

## DESeqAnalysis ====
plotHeatmap(deseq)
plotCorrelationHeatmap(deseq)
plotQuantileHeatmap(deseq)
}
\seealso{
\itemize{
\item \code{pheatmap::pheatmap()}.
\item \code{RColorBrewer::brewer.pal()}.
\item \code{stats::cor()}.
\item \code{stats::hclust()}.
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
