% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotLFC-methods.R
\name{plotLFC}
\alias{plotLFC}
\title{Plot log fold change}
\usage{
plotLFC(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Plot log fold change
}
\note{
Updated 2021-08-03.
}
\examples{
data(deseq)

## Plot the LFC distribution.
plotLFC(deseq, i = 1L)

## Plot expression of specific genes.
dds <- as(deseq, "DESeqDataSet")
genes <- head(rownames(dds))
plotLFC(deseq, i = 1L, genes = genes)
}
