% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMA-methods.R
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DESeqResults-method}
\alias{plotMA,DESeqAnalysis-method}
\alias{plotMeanAverage}
\title{MA-plot: plot differences versus averages for high-throughput data}
\usage{
plotMA(object, ...)

\S4method{plotMA}{DESeqResults}(
  object,
  alphaThreshold = NULL,
  lfcThreshold = NULL,
  baseMeanThreshold = NULL,
  genes = NULL,
  gene2symbol = NULL,
  ntop = 0L,
  direction = c("both", "up", "down"),
  pointColor = c(downregulated = acidplots::lightPalette[["purple"]], upregulated =
    acidplots::lightPalette[["orange"]], nonsignificant =
    acidplots::lightPalette[["gray"]]),
  pointSize = 2L,
  pointAlpha = 0.8
)

\S4method{plotMA}{DESeqAnalysis}(object, i, ...)
}
\arguments{
\item{object}{Object.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{ntop}{\code{integer(1)}.
Number of top features to label.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
MA-plot: plot differences versus averages for high-throughput data
}
\details{
An MA plot is an application of a Bland–Altman plot for visual
representation of genomic data. The plot visualizes the differences between
measurements taken in two samples, by transforming the data onto
M (log ratio) and A (mean average) scales, then plotting these values.
}
\section{Functions}{
\itemize{
\item \code{plotMA,DESeqAnalysis-method}: Passes to \code{DESeqResults} method, with \code{gene2symbol}
argument automatically defined.
}}

\note{
Updated 2020-08-05.
}
\section{plotMA2 aliases}{


Aliased methods for original \code{\link[DESeq2:plotMA]{DESeq2::plotMA()}} S4 methods, which us
geneplotter instead of ggplot2. I prefer using ggplot2 instead, so the
primary methods defined here in the package mask DESeq2.
}

\examples{
data(deseq)

## Get genes from DESeqDataSet.
dds <- as(deseq, "DESeqDataSet")
genes <- head(rownames(dds))
print(genes)

## DESeqAnalysis ====
plotMA(deseq, i = 1L)

## Customize the colors.
plotMA(
    object = deseq,
    i = 1L,
    pointColor = c(
        downregulated = "red",
        nonsignificant = "black",
        upregulated = "green"
    )
)

## Directional support (up or down).
plotMA(deseq, i = 1L, direction = "up", ntop = 5L)
plotMA(deseq, i = 1L, direction = "down", ntop = 5L)

## Label genes manually.
## Note that either gene IDs or names (symbols) are supported.
plotMA(deseq, i = 1L, genes = genes)
}
\seealso{
\code{\link[DESeq2:plotMA]{DESeq2::plotMA()}}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
