% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPCA-methods.R
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,DESeqAnalysis-method}
\title{Plot reduced dimensions}
\usage{
plotPCA(object, ...)

\S4method{plotPCA}{DESeqAnalysis}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot cell-level dimensionality reduction.
}
\details{
Passes to \code{SummarizedExperiment} defined in AcidPlots package.
}
\section{Functions}{
\itemize{
\item \code{plotPCA,DESeqAnalysis-method}: Extracts \code{DESeqTransform}, converts to
\code{RangedSummarizedExperiment}, and passes to method defined in AcidPlots.
}}

\note{
Updated 2022-01-21.
}
\section{Reduction types}{


\itemize{
\item PCA: \strong{P}rincipal \strong{C}omponent \strong{A}nalysis.
\item t-SNE: \strong{t}-distributed \strong{S}tochastic \strong{N}eighbor \strong{E}mbedding.
\item UMAP: \strong{U}niform \strong{M}anifold \strong{A}pproximation and \strong{P}rojection.
}

}

\section{Principal component analysis (\code{plotPCA})}{



PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.

We're using a modified version of the \code{DESeqTransform} method here.\if{html}{\out{<div class="sourceCode r">}}\preformatted{methodFunction(
    f = "plotPCA",
    signature = "DESeqTransform",
    package = "DESeq2"
)
}\if{html}{\out{</div>}}

}

\section{UMAP calculation}{



\href{https://github.com/lmcinnes/umap}{UMAP} calculation in R requires the \href{https://www.python.org}{Python} module \code{umap-learn}.
The UMAP module can be loaded in R using \href{https://rstudio.github.io/reticulate/}{reticulate}.

}

\examples{
data(deseq)

## DESeqAnalysis ====
plotPCA(deseq)
}
\references{
Jolliffe, et al., 2002.
}
\seealso{
\itemize{
\item \code{DESeq2::plotPCA()}.
\item \code{Seurat::DimPlot()}.
\item \code{monocle3::plot_cells()}.
\item \href{https://satijalab.org/seurat/mca.html}{Seurat Mouse Cell Atlas vignette}.
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
