% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTximportFiles.R
\name{prepareTximportFiles}
\alias{prepareTximportFiles}
\title{Prepare quant files for tximport}
\usage{
prepareTximportFiles(
  dir,
  type = c("salmon", "kallisto"),
  makeNames = c("makeNames", "snakeCase", "camelCase")
)
}
\arguments{
\item{dir}{\code{character(1)}.
Directory path containing quant files.
See \code{\link[tximport:tximport]{tximport()}} for details.}

\item{type}{\code{character(1)}.
Type of quant/abundance file to find.
Expecting \code{"quant.sf"} for salmon, \code{"abundance.h5"} for kallisto.}

\item{makeNames}{\code{character(1)}.
Syntactic name function to apply on sample names.
Uses \code{\link[base:match.arg]{match.arg()}} internally.}
}
\value{
\code{character}.
Return quant file paths, with valid sample names automatically applied.
}
\description{
Prepare quant files for tximport
}
\details{
Runs the following internal comments:
\enumerate{
\item Extract sample directory name from quant file using
\code{\link[base:basename]{dirname()}} and \code{\link[base:basename]{basename()}}.
\item Autopad zeros, if necessary, via
\code{\link[basejump:reexports]{autopadZeros()}}.
\item Sanitizes names with \code{\link[syntactic:snakeCase]{snakeCase()}}.
\item Sorts files alphabetically.
}
}
\note{
Updated 2023-04-28.
}
\examples{
library(AcidBase)
tempdir <- tempdir2()
samples <- c("1-sample-A", "2-sample-B")
salmonDir <- initDir(file.path(tempdir, "salmon"))
kallistoDir <- initDir(file.path(tempdir, "kallisto"))
invisible({
    lapply(X = file.path(salmonDir, samples), FUN = initDir)
    file.create(file.path(salmonDir, samples, "quant.sf"))
    lapply(X = file.path(kallistoDir, samples), FUN = initDir)
    file.create(file.path(kallistoDir, samples, "abundance.h5"))
})
salmonFiles <- prepareTximportFiles(dir = salmonDir, type = "salmon")
print(salmonFiles)
kallistoFiles <- prepareTximportFiles(dir = kallistoDir, type = "kallisto")
print(kallistoFiles)
unlink2(tempdir)
}
\seealso{
\itemize{
\item tximport vignette
}
}
