% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/results-methods.R
\name{results}
\alias{results}
\alias{results,DESeqDataSet-method}
\alias{results,DESeqAnalysis-method}
\title{Results}
\usage{
results(object, ...)

\S4method{results}{DESeqDataSet}(object, ...)

\S4method{results}{DESeqAnalysis}(object, i, extra = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{extra}{\code{logical(1)}.
Include row data (i.e. gene metadata) and normalized counts from the
internal \code{DESeqDataSet}.}
}
\value{
\code{DESeqResults}.
}
\description{
Results
}
\note{
Updated 2020-09-21.
}
\section{Extra mode}{


Get the \code{DESeqDataSet}, and humanize the sample names. Note that we're not
calling \code{humanize()} here on the \code{DESeqDataSet}, because we want to keep the
gene identifiers in the row names. Use human-friendly sample names, defined
by the \code{sampleName} column in \code{colData}. We're using this downstream when
joining the normalized counts.
}

\examples{
data(deseq)

## DESeqAnalysis ====
x <- results(deseq, i = 1L)
class(x)
colnames(x)
}
