% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsDiff-methods.R
\name{resultsDiff}
\alias{resultsDiff}
\alias{resultsDiff,DESeqResults,DESeqResults-method}
\alias{resultsDiff,DESeqAnalysis,missingOrNULL-method}
\title{Relative difference of DESeq results}
\usage{
resultsDiff(x, y, ...)

\S4method{resultsDiff}{DESeqResults,DESeqResults}(x, y, col = c("log2FoldChange", "stat"))

\S4method{resultsDiff}{DESeqAnalysis,missingOrNULL}(x, y = NULL, i, j, col = c("log2FoldChange", "stat"))
}
\arguments{
\item{i}{\code{character(1)} or \code{integer(1)}.
\code{DESeqResults} contrast name or position in \code{results} slot.}
}
\description{
Relative difference of DESeq results
}
\note{
Updated 2019-12-17.
}
