% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsNames-methods.R
\name{resultsNames}
\alias{resultsNames}
\alias{resultsNames<-}
\alias{resultsNames,DESeqDataSet-method}
\alias{resultsNames,DESeqAnalysis-method}
\alias{resultsNames<-,DESeqAnalysis,character-method}
\alias{contrastNames}
\alias{contrastNames<-}
\alias{contrastNames,DESeqDataSet-method}
\alias{contrastNames,DESeqAnalysis-method}
\alias{contrastNames<-,DESeqAnalysis,character-method}
\title{Results names}
\usage{
resultsNames(object, ...)

resultsNames(object, ...) <- value

\S4method{resultsNames}{DESeqDataSet}(object)

\S4method{resultsNames}{DESeqAnalysis}(object)

\S4method{resultsNames}{DESeqAnalysis,character}(object) <- value

contrastNames(object, ...)

contrastNames(object, ...) <- value

\S4method{contrastNames}{DESeqDataSet}(object)

\S4method{contrastNames}{DESeqAnalysis}(object)

\S4method{contrastNames}{DESeqAnalysis,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Results names
}
\note{
Updated 2019-09-10.
}
\examples{
data(deseq)
resultsNames(deseq)
}
\seealso{
\itemize{
\item \code{DESeq2::resultsNames()}.
}
}
