% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsTables-methods.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqAnalysis-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqAnalysis}(object, results = 1L,
  rowData = TRUE, counts = TRUE, return = c("tbl_df",
  "DataFrameList"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{character(1)} or \code{integer(1)}. Name or positive of
\code{DESeqResults}.}

\item{rowData}{\code{logical(1)}. Join the row annotations.}

\item{counts}{\code{logical(1)}. Join the size-factor adjusted normalized counts.}

\item{return}{\code{character(1)}. Type of data frame to return in the list. Uses
\code{match.arg}. Note that \code{DataFrame} option will return with rownames,
whereas \code{tbl_df} option will return with \code{"rowname"} column.}
}
\value{
\code{list}. Named list containing subsets of \code{DESeqResults}.
}
\description{
Generate tables summarizing the differential expression, with subsets for
differentially expressed genes (DEGs).
}
\details{
DEG tables (i.e. everything except the \code{all} table), are arranged by adjusted
\emph{P} value.
}
\note{
Do not apply post hoc log fold change cutoffs.
}
\section{Tables}{

\itemize{
\item \code{all}: All genes, including genes without an adjusted \emph{P} value. This table
is unmodified, and the rows have not been re-arranged or subset. It is
suitable for gene set enrichment analysis (GSEA).
\item \code{up}: Upregulated genes.
\item \code{down}: Downregulated genes.
\item \code{both}: Bi-directional DEGs (up- and down-regulated). This table can be
used for overrepresentation testing but should NOT be used for GSEA.
}
}

\examples{
data(deseq)

## DESeqAnalysis ====
x <- resultsTables(deseq)
print(x)
}
