% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sampleData-methods.R
\name{sampleData}
\alias{sampleData}
\alias{sampleData,DESeqAnalysis-method}
\title{Sample data}
\usage{
sampleData(object, ...)

\S4method{sampleData}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Metadata that describes the samples.
}
\details{
This is a complement to the standard
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} function, but improves support
for accessing sample metadata for datasets where multiple items in the
columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\note{
Updated 2021-03-15.
}
\examples{
data(deseq)

## DESeqAnalysis ====
x <- sampleData(deseq)
head(x)
}
