% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables-methods.R
\name{topTables}
\alias{topTables}
\alias{topTables,DESeqResults-method}
\alias{topTables,DESeqAnalysis-method}
\alias{topTables,list-method}
\title{Top tables}
\usage{
topTables(object, ...)

\S4method{topTables}{DESeqResults}(object, DESeqDataSet = NULL, n = 10L)

\S4method{topTables}{DESeqAnalysis}(object, i, n = 10L, lfcShrink = TRUE)

\S4method{topTables}{list}(object, n = 10L, contrast = NULL)
}
\arguments{
\item{object}{Object.}

\item{DESeqDataSet}{\code{DESeqDataSet} or \code{NULL}.}

\item{n}{\code{integer(1)}.
Number of genes (per direction) to report.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}

\item{contrast}{\code{character(1)} or \code{NULL}.
Contrast name.}

\item{...}{Additional arguments.}
}
\value{
\code{kable}.
Markdown tables.
}
\description{
Top tables of differentially expressed genes.
}
\section{Functions}{
\itemize{
\item \code{topTables,list-method}: Legacy support for \code{tbl_df} list returned from
\code{resultsTables()}. This method is still supported because it is used in the
F1000 v2 workflow paper. Otherwise, we now recommend using the
\code{DESeqAnalysis} method directly.
}}

\note{
Updated 2019-11-12.
}
\examples{
data(deseq)

## DESeqAnalysis ====
topTables(deseq, i = 1L, n = 5L, lfcShrink = TRUE)

## DESeqResults 'resultsTables()' list ====
res <- results(deseq, i = 1L, lfcShrink = TRUE)
resTbl <- resultsTables(res, return = "tbl_df")
topTables(resTbl)
}
