% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSummary-methods.R
\name{alphaSummary}
\alias{alphaSummary}
\alias{alphaSummary,DESeqDataSet-method}
\title{Alpha level cutoff summary statistics}
\usage{
alphaSummary(object, ...)

\S4method{alphaSummary}{DESeqDataSet}(
  object,
  alpha = c(0.1, 0.05, 0.01, 0.001, 1e-06),
  contrast = NULL,
  name = NULL
)
}
\arguments{
\item{object}{Object.}

\item{alpha}{\code{numeric}. Multiple alpha cutoffs.}

\item{contrast}{\code{character}.
A character vector with exactly 3 elements:
\enumerate{
\item Name of factor in the design formula.
\item Name of numerator level for the fold change.
\item Name of denominator level for the fold change.
}

See \code{\link[DESeq2:results]{DESeq2::results()}} for details.}

\item{name}{\code{character(1)}.
Name of the individual effect (coefficient) for building a results table.
Use this argument rather than \code{contrast} for continuous variables.}

\item{...}{Additional arguments.}
}
\value{
\verb{integer matrix}.
}
\description{
Quickly generate a summary table of various alpha level cutoffs.
}
\details{
Use either \code{contrast} or \code{name} to specify the desired contrast.
}
\note{
Updated 2019-08-20.
}
\examples{
data(deseq)

## DESeqDataSet ====
dds <- as(deseq, "DESeqDataSet")
resultsNames(dds)
alphaSummary(dds)
alphaSummary(dds, contrast = c("condition", "B", "A"))
alphaSummary(dds, name = "condition_B_vs_A")
}
\seealso{
\itemize{
\item \code{DESeq2::results()}.
\item \code{DESeq2::resultsNames()}.
}
}
\author{
Michael Steinbaugh, Lorena Patano
}
