% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/alphaThreshold-methods.R
\name{alphaThreshold}
\alias{alphaThreshold}
\alias{alphaThreshold<-}
\alias{alphaThreshold,DESeqResults-method}
\alias{alphaThreshold,DESeqAnalysis-method}
\alias{alphaThreshold<-,DESeqAnalysis,numeric-method}
\alias{alphaThreshold<-,DESeqAnalysis,NULL-method}
\title{Alpha threshold}
\usage{
alphaThreshold(object, ...)

alphaThreshold(object, ...) <- value

\S4method{alphaThreshold}{DESeqResults}(object)

\S4method{alphaThreshold}{DESeqAnalysis}(object)

\S4method{alphaThreshold}{DESeqAnalysis,numeric}(object) <- value

\S4method{alphaThreshold}{DESeqAnalysis,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)}.
Alpha cutoff level.
}
\description{
Alpha threshold
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
object <- deseq
alphaThreshold(object) <- 0.05
alphaThreshold(object)
}
