% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R
\name{as}
\alias{as}
\alias{coerce}
\alias{coerce,DESeqAnalysis,DESeqDataSet-method}
\alias{coerce,DESeqAnalysis,DESeqTransform-method}
\alias{coerce,DESeqAnalysis,DESeqResults-method}
\title{Force an object to belong to a class}
\description{
Force an object to belong to a class
}
\note{
Updated 2019-11-12.
}
\section{DESeqAnalysis}{


Supported coercion methods will extract any of these internal objects:
\itemize{
\item \code{DESeqDataSet}.
\item \code{DESeqTransform}.
\item \code{DESeqResults}. Extracts the first results slotted. Note that this
corresponds to results containing log2 fold change (LFC) values that \emph{have
not been shrunken} using \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.
}
}

\examples{
data(deseq)

## DESeqAnalysis ====
dds <- as(deseq, "DESeqDataSet")
print(dds)
dt <- as(deseq, "DESeqTransform")
print(dt)
## Pulls the first results slotted.
res <- as(deseq, "DESeqResults")
summary(res)
}
\author{
Michael Steinbaugh
}
