% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R
\name{as}
\alias{as}
\alias{coerce}
\alias{coerce,DESeqAnalysis,DESeqDataSet-method}
\alias{coerce,DESeqAnalysis,DESeqTransform-method}
\alias{coerce,DESeqAnalysis,DESeqResults-method}
\title{Force an object to belong to a class}
\description{
Coerce an object to a given class.
}
\note{
Updated 2019-08-20.
}
\section{DESeqAnalysis}{


Supported coercion methods will extract any of these internal objects:
\itemize{
\item \code{DESeqDataSet}.
\item \code{DESeqTransform}.
\item \code{DESeqResults}. Extracts the first results slotted. Note that this
corresponds to results containing log2 fold change (LFC) values that \emph{have
not been shrunken} using \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.
}
}

\section{Description}{

    \code{as(object)}
      returns the version of this object coerced to be the given
      \code{Class}.  When used in the replacement form on the left of
      an assignment, the portion of the object corresponding to
      \code{Class} is replaced by \code{value}.

      The operation of \code{as()} in either form depends on the
      definition of coerce methods.  Methods are defined automatically
      when the two classes are related by inheritance; that is, when
      one of the classes is a subclass of the other.

      Coerce methods are also predefined for basic classes (including all
      the types of vectors, functions and a few others).

      Beyond these two sources of methods, further methods are defined
      by calls to the \code{\link{setAs}} function.  See that
      documentation also for details of how coerce methods work. Use
      \code{showMethods(coerce)} for a list of all currently defined methods, as in the
      example below.

 
  
}

\section{Basic Coercion Methods}{

  Methods are pre-defined for coercing any object to one of the basic
  datatypes.  For example, \code{as(x, "numeric")} uses the existing
  \code{as.numeric} function.  These and all other existing methods
  can be listed as shown in the example.

}

\examples{
data(deseq)

## DESeqAnalysis ====
dds <- as(deseq, "DESeqDataSet")
print(dds)
dt <- as(deseq, "DESeqTransform")
print(dt)
## Pulls the first results slotted.
res <- as(deseq, "DESeqResults")
contrastName(res)
summary(res)
}
\references{
Chambers, John M. (2016)
 \emph{Extending R},
  Chapman & Hall.
(Chapters 9 and 10.)
}
\seealso{
If you think of using \code{try(as(x, cl))}, consider
  \code{\link{canCoerce}(x, cl)} instead.
}
\author{
Michael Steinbaugh
}
