% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/baseMeanThreshold-methods.R
\name{baseMeanThreshold}
\alias{baseMeanThreshold}
\alias{baseMeanThreshold<-}
\alias{baseMeanThreshold,DESeqAnalysis-method}
\alias{baseMeanThreshold,DESeqAnalysisList-method}
\alias{baseMeanThreshold,DESeqResults-method}
\alias{baseMeanThreshold,DESeqResultsList-method}
\alias{baseMeanThreshold<-,DESeqAnalysis,numeric-method}
\alias{baseMeanThreshold<-,DESeqAnalysis,NULL-method}
\alias{baseMeanThreshold<-,DESeqAnalysisList,numeric-method}
\alias{baseMeanThreshold<-,DESeqAnalysisList,NULL-method}
\alias{baseMeanThreshold<-,DESeqResults,numeric-method}
\alias{baseMeanThreshold<-,DESeqResults,NULL-method}
\alias{baseMeanThreshold<-,DESeqResultsList,numeric-method}
\alias{baseMeanThreshold<-,DESeqResultsList,NULL-method}
\title{Base mean threshold}
\usage{
baseMeanThreshold(object, ...)
baseMeanThreshold(object, ...) <- value

\S4method{baseMeanThreshold}{DESeqAnalysis}(object)

\S4method{baseMeanThreshold}{DESeqAnalysisList}(object)

\S4method{baseMeanThreshold}{DESeqResults}(object)

\S4method{baseMeanThreshold}{DESeqResultsList}(object)

\S4method{baseMeanThreshold}{DESeqAnalysis,numeric}(object) <- value

\S4method{baseMeanThreshold}{DESeqAnalysis,`NULL`}(object) <- value

\S4method{baseMeanThreshold}{DESeqAnalysisList,numeric}(object) <- value

\S4method{baseMeanThreshold}{DESeqAnalysisList,`NULL`}(object) <- value

\S4method{baseMeanThreshold}{DESeqResults,numeric}(object) <- value

\S4method{baseMeanThreshold}{DESeqResults,`NULL`}(object) <- value

\S4method{baseMeanThreshold}{DESeqResultsList,numeric}(object) <- value

\S4method{baseMeanThreshold}{DESeqResultsList,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)}.
Base mean cutoff level.
}
\description{
Base mean threshold
}
\note{
Updated 2021-03-12.
}
\examples{
data(deseq)

## DESeqAnalysis ====
object <- deseq
baseMeanThreshold(object) <- 100L
baseMeanThreshold(object)
}
