% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastName-methods.R
\docType{methods}
\name{contrastName}
\alias{contrastName}
\alias{contrastName<-}
\alias{contrastName,DESeqResults-method}
\alias{contrastName<-,DESeqResults,character-method}
\alias{contrastName,DESeqAnalysis-method}
\title{Contrast name}
\usage{
contrastName(object, ...)

contrastName(object, ...) <- value

\S4method{contrastName}{DESeqResults}(object, format = c("resultsNames",
  "title"), useStash = TRUE)

\S4method{contrastName}{DESeqResults,character}(object) <- value

\S4method{contrastName}{DESeqAnalysis}(object, i, ...)
}
\arguments{
\item{object}{Object.}

\item{format}{\code{character(1)}.
Name format to return:
\itemize{
\item \code{resultsNames}: Attempt to matching the conventions in
\code{\link[DESeq2:resultsNames]{resultsNames()}}.
\item \code{title}: Human readable, for plot titles and/or table captions.
}}

\item{useStash}{\code{logical(1)}.
Check for \code{contrastName} metadata stash in \code{DESeqResults} object. Intended
for use with \code{DESeqAnalysis} methods.}

\item{value}{Value to assign.}

\item{i}{\code{integer(1)} or \code{character(1)}.
Position or name corresponding to \code{DESeqResults} slotted in \code{results}.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Contrast name.
}
\description{
Contrast name
}
\note{
Updated 2019-11-08.
}
\examples{
data(deseq)

## DESeqResults ====
object <- as(deseq, "DESeqResults")
contrastName(object)

## DESeqAnalysis ====
contrastName(deseq, i = 1L)
}
\seealso{
\itemize{
\item \code{\link[=contrastNames]{contrastNames()}}.
\item \code{\link[DESeq2:resultsNames]{resultsNames()}}.
}
}
