% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastName-methods.R
\docType{methods}
\name{contrastName}
\alias{contrastName}
\alias{contrastName,DESeqResults-method}
\alias{contrastName,DESeqAnalysis-method}
\title{Contrast name}
\usage{
contrastName(object, ...)

\S4method{contrastName}{DESeqResults}(object, format = c("resultsNames",
  "title"))

\S4method{contrastName}{DESeqAnalysis}(object, results)
}
\arguments{
\item{object}{Object.}

\item{format}{\code{character(1)}.
Name format to return:
\itemize{
\item \code{resultsNames}: Attempt to matching the conventions in
\code{\link[DESeq2:resultsNames]{resultsNames()}}.
\item \code{title}: Human readable, for plot titles and/or table captions.
}}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}
}
\value{
\code{character(1)}.
Contrast name.
}
\description{
Contrast name
}
\examples{
data(deseq)

## DESeqResults ====
object <- as(deseq, "DESeqResults")
contrastName(object)

## DESeqAnalysis ====
contrastName(deseq, results = 1L)
}
\seealso{
\code{\link[DESeq2:resultsNames]{resultsNames()}}.
}
