% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastName-methods.R
\name{contrastName}
\alias{contrastName}
\alias{contrastName<-}
\alias{contrastName,DESeqAnalysis-method}
\alias{contrastName,DESeqResults-method}
\alias{contrastName<-,DESeqResults,character-method}
\title{Contrast name}
\usage{
contrastName(object, ...)
contrastName(object, ...) <- value

\S4method{contrastName}{DESeqAnalysis}(object, i, ...)

\S4method{contrastName}{DESeqResults}(object, format = c("resultsNames", "title"))

\S4method{contrastName}{DESeqResults,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{...}{Additional arguments.}

\item{format}{\code{character(1)}.
Name format to return:
\itemize{
\item \code{resultsNames}: Attempt to matching the conventions in
\code{\link[DESeq2:results]{resultsNames()}}.
\item \code{title}: Human readable, for plot titles and/or table captions.
}}

\item{value}{Value to assign.}
}
\value{
\code{character(1)}.
Contrast name.
}
\description{
Contrast name
}
\note{
Updated 2021-03-15.
}
\examples{
data(deseq)

## DESeqAnalysis ====
contrastName(deseq, i = 1L)
}
\seealso{
\itemize{
\item \code{\link[=resultsNames]{resultsNames()}}.
}
}
