% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastName-methods.R
\docType{methods}
\name{contrastName}
\alias{contrastName}
\alias{contrastName,DESeqResults-method}
\alias{contrastName,DESeqAnalysis-method}
\title{Contrast name}
\usage{
\S4method{contrastName}{DESeqResults}(object)

\S4method{contrastName}{DESeqAnalysis}(object, results)
}
\arguments{
\item{object}{Object.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}
}
\value{
\code{character(1)}.
Contrast name.
}
\description{
Contrast name
}
\examples{
data(deseq)

## DESeqResults ====
object <- as(deseq, "DESeqResults")
contrastName(object)

## DESeqAnalysis ====
contrastName(deseq, results = 1L)
}
