% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastSamples-methods.R
\name{contrastSamples}
\alias{contrastSamples}
\alias{contrastSamples,DESeqAnalysis-method}
\title{Samples corresponding to a differential expression contrast}
\usage{
contrastSamples(object, ...)

\S4method{contrastSamples}{DESeqAnalysis}(object, i, ...)
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Sample identifiers.
}
\description{
Samples corresponding to a differential expression contrast
}
\details{
Match the samples in a \code{DESeqDataSet} used to define contrast in a
corresponding \code{DESeqResults} object. Note that this only works for simple
(e.g. pairwise) contrasts and will intentionally error for more complex
comparisons.
}
\note{
Updated 2020-08-04.

\strong{EXPERIMENTAL}. This approach will error intentionally for
\code{DESeqResults} objects generated with complex contrasts (e.g. interaction
effect or LRT). It may be removed in a future release if this approach is
too error-prone.
}
\examples{
data(deseq)

## DESeqAnalysis ====
contrastSamples(deseq, i = 1L)
}
