% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastSamples-methods.R
\name{contrastSamples}
\alias{contrastSamples}
\alias{contrastSamples,DESeqAnalysis-method}
\title{Samples corresponding to a differential expression contrast}
\usage{
contrastSamples(object, ...)

\S4method{contrastSamples}{DESeqAnalysis}(object, i, quiet = FALSE, return = c("character", "list"))
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{character}: Sample identifiers, corresponding to the column names
of \code{DESeqDataSet}.
\item \code{list}: Named list containing \code{"contrast"} and \code{"samples"} elements:
(1) \code{"contrast"}: \code{character} vector containing metadata on \code{"factor"},
\code{"numerator"}, and \code{"denominator"} contrast elements;
(2) \code{"samples"}: ``list\code{containing}"numerator"` and `"denominator"` of
sample identifiers corresponding to column names of `DESeqDataSet`.
}
}
\description{
Samples corresponding to a differential expression contrast
}
\details{
Match the samples in a \code{DESeqDataSet} used to define contrast in a
corresponding \code{DESeqResults} object. Note that this only works for simple
(e.g. pairwise) contrasts and will intentionally error for more complex
comparisons.
}
\note{
Updated 2022-05-17.
}
\examples{
data(deseq)

## DESeqAnalysis ====
x <- contrastSamples(deseq, i = 1L, return = "list")
print(x)

x <- contrastSamples(deseq, i = 1L, return = "character")
print(x)
}
