% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/correlation-methods.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(x, y, i, j, ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{i}{\code{integer(1)} or \code{character(1)}.
Primary assay.}

\item{j}{\code{integer(1)}, \code{character(1)}, or \code{NULL}.
Optional secondary assay.
If \code{NULL}, calculates correlation matrix only on the primary assay.}

\item{...}{Additional arguments.}
}
\description{
Correlation
}
\note{
Updated 2019-11-07.
}
\examples{
data(deseq)
dds <- as(deseq, "DESeqDataSet")
correlation(dds)
}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{S4Vectors::cor()}.
}
}
