% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/correlation-methods.R
\docType{methods}
\name{correlation}
\alias{correlation}
\alias{correlation,matrix-method}
\alias{correlation,Matrix-method}
\title{Correlation}
\usage{
correlation(x, ...)

\S4method{correlation}{matrix}(x, y = NULL, method = c("pearson",
  "kendall", "spearman"))

\S4method{correlation}{Matrix}(x, y = NULL, method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{y}{Object.}

\item{i}{\code{integer(1)} or \code{character(1)}.
Primary assay.}

\item{j}{\code{integer(1)}, \code{character(1)}, or \code{NULL}.
Optional secondary assay.
If \code{NULL}, calculates correlation matrix only on the primary assay.}
}
\description{
Correlation
}
\note{
Updated 2019-11-07.
}
\examples{
data(deseq)
dds <- as(deseq, "DESeqDataSet")
correlation(dds)
}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{stats::cor.test()}.
\item \code{stats::sd()}.
\item \code{S4Vectors::cor()}.
\item https://stats.stackexchange.com/questions/24980
}
}
