% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-methods.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(x, y, ...)
}
\value{
\code{numeric(1)} or \code{matrix}.
}
\description{
Correlation
}
\note{
Updated 2019-11-08.
}
\examples{
data(correlation, package = "acidtest")
list <- correlation
rm(correlation)

## DESeqResults ====
x <- list[["vector_x"]]
y <- list[["vector_y"]]

head(x)
head(y)

correlation(x = x, y = y)

## DESeqAnalysis ====
x <- list[["matrix_x"]]
y <- list[["matrix_y"]]

head(x)
head(y)

stats::cor(x)
correlation(x)

stats::cor(x = c(x), y = c(y))
correlation(x = x, y = y)

## SummarizedExperiment ====
x <- list[["SummarizedExperiment_x"]]
y <- list[["SummarizedExperiment_y"]]

correlation(x = x, i = 1L)
correlation(x = x, i = 1L, j = 2L)
correlation(x = x, y = y)
}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{stats::cor.test()}.
\item \code{stats::sd()}.
\item \code{S4Vectors::cor()}.
\item https://stats.stackexchange.com/questions/24980
}
}
