% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-methods.R
\name{correlation}
\alias{correlation}
\alias{correlation,DESeqResults,DESeqResults-method}
\alias{correlation,DESeqAnalysis,missingOrNULL-method}
\title{Correlation}
\usage{
correlation(x, y, ...)

\S4method{correlation}{DESeqResults,DESeqResults}(
  x,
  y,
  col = "log2FoldChange",
  method = c("pearson", "kendall", "spearman")
)

\S4method{correlation}{DESeqAnalysis,missingOrNULL}(
  x,
  y = NULL,
  i,
  j,
  col = "log2FoldChange",
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{col}{\code{character(1)}.
Column name.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{i}{\code{integer(1)} or \code{character(1)}.
For \code{SummarizedExperiment}, primary assay.}

\item{j}{\code{integer(1)}, \code{character(1)}, or \code{NULL}.
For \code{SummarizedExperiment}, optional secondary assay.
If \code{NULL}, calculates correlation matrix only on the primary assay.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)} or \code{matrix}.
}
\description{
Correlation
}
\note{
Updated 2021-02-09.
}
\examples{
data(deseq)

## DESeqAnalysis ====
correlation(deseq, i = 1L, j = 2L)
}
\seealso{
\itemize{
\item \code{stats::cor()}.
\item \code{stats::cor.test()}.
\item \code{stats::sd()}.
\item \code{S4Vectors::cor()}.
\item https://stats.stackexchange.com/questions/24980
}
}
