% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deg-methods.R
\name{deg}
\alias{deg}
\alias{deg,DESeqAnalysis-method}
\alias{deg,DESeqResults-method}
\title{Differentially expressed genes}
\usage{
deg(object, ...)

\S4method{deg}{DESeqAnalysis}(object, i, direction = c("both", "up", "down"), quiet = FALSE)

\S4method{deg}{DESeqResults}(
  object,
  direction = c("both", "up", "down"),
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  quiet = FALSE
)
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Gene identifiers.
}
\description{
Differentially expressed genes
}
\note{
Updated 2020-08-09.
}
\examples{
data(deseq)

## DESeqAnalysis ====
x <- deg(deseq, i = 1L)
head(x)
}
