% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg-methods.R
\docType{methods}
\name{deg}
\alias{deg}
\alias{deg,DESeqResults-method}
\alias{deg,DESeqAnalysis-method}
\title{Differentially expressed genes}
\usage{
\S4method{deg}{DESeqResults}(object, alpha = NULL, lfcThreshold = NULL,
  direction = c("both", "up", "down"))

\S4method{deg}{DESeqAnalysis}(object, results, direction = c("both",
  "up", "down"))
}
\arguments{
\item{object}{Object.}

\item{alpha}{\code{numeric(1)}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)}.
Log fold change ratio (base 2) cutoff threshold. If left \code{NULL}, will use
the cutoff defined in the object.}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{results}{\code{character(1)} or \code{integer(1)}. Name or positive of
\code{DESeqResults}.}
}
\value{
\code{character}.
Gene identifiers.
}
\description{
Differentially expressed genes
}
\examples{
data(deseq)

## DESeqAnalysis ====
deg(deseq, results = 1L)
}
