% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg-methods.R
\docType{methods}
\name{deg}
\alias{deg}
\alias{deg,DESeqResults-method}
\alias{deg,DESeqAnalysis-method}
\title{Differentially expressed genes}
\usage{
deg(object, ...)

\S4method{deg}{DESeqResults}(object, alpha = NULL, lfcThreshold = NULL,
  direction = c("both", "up", "down"))

\S4method{deg}{DESeqAnalysis}(object, i, direction = c("both", "up",
  "down"), ...)
}
\arguments{
\item{object}{Object.}

\item{alpha}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log fold change ratio (base 2) cutoff threshold. If left \code{NULL}, will use
the cutoff defined in the object.}

\item{direction}{\code{character(1)}.
Plot \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Gene identifiers.
}
\description{
Differentially expressed genes
}
\note{
Updated 2019-11-08.
}
\examples{
data(deseq)

## DESeqAnalysis ====
x <- deg(deseq, i = 1L)
head(x)
}
