% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/degIntersection-methods.R
\name{degIntersection}
\alias{degIntersection}
\alias{degIntersection,DESeqAnalysis-method}
\alias{degIntersection,DESeqAnalysisList-method}
\alias{degIntersection,DESeqResultsList-method}
\title{Intersection of differentially expressed genes}
\usage{
degIntersection(object, ...)

\S4method{degIntersection}{DESeqAnalysis}(
  object,
  i = NULL,
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  direction = c("up", "down"),
  return = c("matrix", "count", "ratio", "names")
)

\S4method{degIntersection}{DESeqAnalysisList}(
  object,
  i = NULL,
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  direction = c("up", "down"),
  return = c("matrix", "count", "ratio", "names")
)

\S4method{degIntersection}{DESeqResultsList}(
  object,
  direction = c("up", "down"),
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  return = c("matrix", "count", "ratio", "names")
)
}
\arguments{
\item{object}{Object.}

\item{i}{\code{character}, \code{numeric}, or \code{NULL}.
Names or range of results.
If set \code{NULL}, include all results per object.
When passing in multiple objects, specify the desired results as a \code{list}
with length matching the number of input objects, containing either
\code{character} or \code{numeric} corresponding to each object.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{direction}{\code{character(1)}.
Include "up" or "down" directions.
Must be directional, and intentionally does not support "both".}

\item{return}{\code{character(1)}.
\itemize{
\item \code{"matrix"}: \verb{logical matrix};
Intersection matrix.
\item \code{"count"}: \code{integer};
Number of times the gene is significant across contrasts.
\item \code{"ratio"}: \code{numeric};
The ratio of how many times the gene is significant across contrasts.
\item \code{"names"}: \code{character};
Names of genes that intersect across all contrasts defined.
Input of specific contrasts with \code{i} argument is recommended here.
}}

\item{...}{Passthrough arguments to \code{DESeqResultsList} method.}
}
\value{
\code{integer}.
Integer denoting the number of intersections, sorted from highest to
lowest. Gene identifiers are defined as the names.
}
\description{
Intersection of differentially expressed genes
}
\note{
Updated 2022-05-17.
}
\examples{
data(deseq)

## DESeqAnalysis ====
mat <- degIntersection(deseq, return = "matrix")
class(mat)
type(mat)
head(mat)

count <- degIntersection(deseq, return = "count")
class(count)
head(count)

ratio <- degIntersection(deseq, return = "ratio")
class(ratio)
head(ratio)

names <- degIntersection(deseq, i = c(1L, 2L), return = "names")
class(names)
head(names)

## DESeqAnalysisList ====
object <- DESeqAnalysisList(list("object1" = deseq, "object2" = deseq))
print(object)
}
