% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degIntersection-methods.R
\name{degIntersection}
\alias{degIntersection}
\alias{degIntersection,DESeqAnalysis-method}
\title{Intersection of differentially expressed genes}
\usage{
degIntersection(object, ...)

\S4method{degIntersection}{DESeqAnalysis}(object, i = NULL, return = c("matrix", "count", "ratio"), ...)
}
\arguments{
\item{object}{Object.}

\item{i}{\code{character}, \code{numeric}, or \code{NULL}.
Names or range of results.
If set \code{NULL}, include all results.}

\item{return}{\code{character(1)}.
Return intersection matrix, count per contrast, or ratio.}

\item{...}{Passthrough arguments to \code{\link[=deg]{deg()}}.}
}
\value{
\code{integer}.
Integer denoting the number of intersections, sorted from highest to
lowest. Gene identifiers are defined as the names.
}
\description{
Intersection of differentially expressed genes
}
\note{
Updated 2020-08-18.
}
\examples{
data(deseq)

## DESeqAnalysis ====
x <- degIntersection(deseq)
head(x)
}
