% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/degPerContrast-methods.R
\name{degPerContrast}
\alias{degPerContrast}
\alias{degPerContrast,DESeqAnalysis-method}
\title{Summary of differentially expressed genes per contrast}
\usage{
degPerContrast(object, ...)

\S4method{degPerContrast}{DESeqAnalysis}(
  object,
  direction = c("both", "up", "down"),
  n = FALSE,
  return = c("matrix", "list"),
  ...
)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=deg]{deg()}}.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}
}
\value{
\itemize{
\item \code{matrix}: Matrix containing numbers of DEGs per contrast.
Intended primarily for use with \code{\link[=plotDEGStackedBar]{plotDEGStackedBar()}}.
\item \code{list}: Named vector containing the DEG identifiers.
Intended primarily for use with \code{\link[=plotDEGUpset]{plotDEGUpset()}}.
}
}
\description{
Summary of differentially expressed genes per contrast
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
degPerContrast(deseq)
}
