% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/degPerContrast-methods.R
\name{degPerContrast}
\alias{degPerContrast}
\alias{degPerContrast,DESeqAnalysis-method}
\title{Base mean threshold}
\usage{
degPerContrast(object, ...)

\S4method{degPerContrast}{DESeqAnalysis}(
  object,
  i = NULL,
  direction = c("both", "up", "down"),
  return = c("matrix", "list"),
  ...
)
}
\arguments{
\item{object}{Object.}

\item{i}{\code{character}, \code{numeric}, or \code{NULL}.
Names or range of results.
If set \code{NULL}, include all results.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Passthrough arguments to \code{deg()}.}
}
\value{
\itemize{
\item \code{matrix}: Matrix containing numbers of DEGs per contrast.
Intended primarily for use with \code{plotDEGStackedBar()}.
\item \code{list}: Named vector containing the DEG identifiers.
Intended primarily for use with \code{plotDEGUpset()}.
}
}
\description{
Base mean threshold
}
\note{
Updated 2020-09-21.
}
\examples{
data(deseq)

## DESeqAnalysis ====
degPerContrast(deseq)
}
