% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deseq}
\alias{deseq}
\title{DESeq2 example analysis}
\format{
An object of class \code{DESeqAnalysis} with 500 rows and 12 columns.
}
\usage{
data(deseq)
}
\description{
DESeq2 example analysis
}
\note{
Updated 2022-05-17.
}
\examples{
data(deseq)
print(deseq)
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
