% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DESeqAnalysis.R
\docType{class}
\name{DESeqAnalysis-class}
\alias{DESeqAnalysis-class}
\alias{DESeqAnalysis}
\title{DESeq2 differential expression analysis}
\usage{
DESeqAnalysis(data, transform, results, lfcShrink = NULL)
}
\arguments{
\item{data}{\code{DESeqDataSet}.}

\item{transform}{\code{DESeqTransform}.
\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}} recommended by default.}

\item{results}{\code{list} or single \code{DESeqResults}.
One or more unshrunken \code{DESeqResults}.
Assign the \code{\link[DESeq2:results]{DESeq2::results()}} return here.}

\item{lfcShrink}{\code{list}, single \code{DESeqResults}, or \code{NULL}.
\emph{Optional}. One or more shrunken \code{DESeqResults}.
Assign the \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return here.}
}
\value{
\code{DESeqAnalysis}.
Contains a \code{DESeqDataSet}, \code{DESeqTransform}, and corresponding
\code{DESeqResults} list.
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2. This class is essentially a \code{list} with validity checks
to ensure \code{DESeqTransform} and \code{DESeqResults} correspond to the
\code{DESeqDataSet}.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{DESeqDataSet}.}

\item{\code{transform}}{\code{DESeqTransform}.}

\item{\code{results}}{\code{list}.
One or more unshrunken \code{DESeqResults}.}

\item{\code{lfcShrink}}{\code{list}.
\emph{Optional}. One or more shrunken \code{DESeqResults}. If set, must correspond to
those defined in \code{results}. Otherwise, can set as empty list (\code{list()}).}
}}

\section{DESeqDataSet}{


We recommend generating the \code{DESeqDataSet} by coercion from \code{bcbioRNASeq}
object using \code{as(dds, "bcbioRNASeq")}. Don't use the \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}
or \code{\link[DESeq2:DESeqDataSetFromMatrix]{DESeq2::DESeqDataSetFromMatrix()}} constructors to generate the
\code{DESeqDataSet} object.
}

\section{DESeqTransform}{


Object containing variance-stabilized counts. We recommend slotting the
return from either \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}} or
\code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
}

\section{DESeqResults}{


Don't modify any of the \code{DESeqResults} objects manually. This includes
rearranging the rows or dropping genes without adjusted P values. We'll take
care of this automatically in supported methods.
}

\examples{
library(DESeq2)

data <- DESeq(makeExampleDESeqDataSet())
class(data)

transform <- varianceStabilizingTransformation(data)
class(transform)

resultsNames(data)
name <- resultsNames(data)[[2L]]
results <- results(data, name = name)
class(results)

lfcShrink <- lfcShrink(dds = data, res = results, coef = 2L)

results <- list(results)
names(results) <- name

lfcShrink <- list(lfcShrink)
names(lfcShrink) <- name

identical(names(results), names(lfcShrink))

x <- DESeqAnalysis(
    data = data,
    transform = transform,
    results = results,
    lfcShrink = lfcShrink
)
print(x)
}
\seealso{
\itemize{
\item \code{help(topic = "Annotated-class", package = "S4Vectors")}.
\item \code{showClass("Annotated")}.
}
}
\author{
Michael Steinbaugh
}
