% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{DESeqAnalysis-package}
\title{DESeqAnalysis}
\description{
Toolkit for performing differential expression with
\href{https://bioconductor.org/packages/DESeq2/}{DESeq2}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/deseqanalysis/}
  \item \url{https://github.com/acidgenomics/r-deseqanalysis/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-deseqanalysis/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
