% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqAnalysis.R
\name{DESeqAnalysis}
\alias{DESeqAnalysis}
\title{DESeq2 Analysis Container}
\usage{
DESeqAnalysis(data, transform, results, lfcShrink)
}
\arguments{
\item{data}{\code{DESeqDataSet}.}

\item{transform}{\code{DESeqTransform}.}

\item{results}{\code{list}. One or more unshrunken \code{DESeqResults}. Assign the
\code{DESeq2::results} return here.}

\item{lfcShrink}{\code{list}. One or more shrunken \code{DESeqResults}. Assign the
\code{DESeq2::lfcShrink} return here.}
}
\value{
\code{DESeqAnalysis}, which contains a \code{DESeqDataSet}, \code{DESeqTransform},
and corresponding \code{DESeqResults}.
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2. This class is essentially a \code{list} with validity checks
to ensure \code{DESeqTransform} and \code{DESeqResults} correspond to the
\code{DESeqDataSet}.
}
\section{DESeqDataSet}{


We recommend generating the \code{DESeqDataSet} by coercion from \code{bcbioRNASeq}
object using \code{as(dds, "bcbioRNASeq")}. Don't use the \code{DESeq2::DESeqDataSet}
or \code{DESeq2::DESeqDataSetFromMatrix} constructors to generate the
\code{DESeqDataSet} object.
}

\section{DESeqResults}{


Don't modify any of the \code{DESeqResults} objects manually. This includes
rearranging the rows or dropping genes without adjusted P values. We'll take
care of this automatically in supported methods.
}

\examples{
library(DESeq2)
dds <- DESeq(makeExampleDESeqDataSet())
class(dds)

dt <- varianceStabilizingTransformation(dds)
class(dt)

resultsNames(dds)
res <- results(dds, name = resultsNames(dds)[[2L]])
class(res)

x <- DESeqAnalysis(
    data = dds,
    transform = dt,
    results = list(res),
    lfcShrink = list(lfcShrink(dds = dds, res = res, coef = 2L))
)
print(x)
}
\seealso{
\code{DESeqAnalysis}.
}
