% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{DESeqAnalysis}
\alias{DESeqAnalysis}
\title{DESeq2 differential expression analysis}
\usage{
DESeqAnalysis(data, transform, results, lfcShrink = NULL)
}
\arguments{
\item{data}{\code{DESeqDataSet}.}

\item{transform}{\code{DESeqTransform}.
\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}} recommended by default.}

\item{results}{\code{list} or single \code{DESeqResults}.
One or more unshrunken \code{DESeqResults}.
Assign the \code{\link[DESeq2:results]{DESeq2::results()}} return here.}

\item{lfcShrink}{\code{list}, single \code{DESeqResults}, or \code{NULL}.
\emph{Optional}. One or more shrunken \code{DESeqResults}.
Assign the \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return here.}
}
\value{
\code{DESeqAnalysis}.
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2. This class is essentially a \code{list} with validity checks
to ensure \code{DESeqTransform} and \code{DESeqResults} correspond to the
\code{DESeqDataSet}.
}
\note{
Updated 2019-08-20.
}
\examples{
suppressPackageStartupMessages(library(DESeq2))

data <- makeExampleDESeqDataSet()
data <- DESeq(data)
class(data)

## Gene identifiers and names (symbols) are required.
rowRanges <- emptyRanges(names = rownames(data))
mcols(rowRanges)[["geneID"]] <- paste0("id", seq_len(length(rowRanges)))
mcols(rowRanges)[["geneName"]] <- paste0("name", seq_len(length(rowRanges)))

transform <- varianceStabilizingTransformation(data)
class(transform)

resultsNames(data)
name <- resultsNames(data)[[2L]]
results <- results(data, name = name)
class(results)

lfcShrink <- lfcShrink(dds = data, res = results, coef = 2L)

results <- list(results)
names(results) <- name

lfcShrink <- list(lfcShrink)
names(lfcShrink) <- name

object <- DESeqAnalysis(
    data = data,
    transform = transform,
    results = results,
    lfcShrink = lfcShrink
)
print(object)
}
