% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqAnalysisList-class}
\alias{DESeqAnalysisList-class}
\title{List containing related DESeq2 analyses}
\value{
\code{DESeqAnalysisList}.
}
\description{
List containing related DESeq2 analyses
}
\note{
Updated 2021-03-09.
}
\author{
Michael Steinbaugh
}
