% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DESeqAnalysisList.R
\docType{class}
\name{DESeqAnalysisList-class}
\alias{DESeqAnalysisList-class}
\alias{DESeqAnalysisList}
\title{List containing related DESeq2 analyses}
\usage{
DESeqAnalysisList(...)
}
\arguments{
\item{...}{\code{DESeqAnalysis} objects.}
}
\description{
List containing related DESeq2 analyses
}
\examples{
data(deseq)
x <- DESeqAnalysisList(deseq)
x
}
\author{
Michael Steinbaugh
}
