% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqResultsList-class}
\alias{DESeqResultsList-class}
\title{List containing related DESeqResults objects}
\value{
\code{DESeqResultsList}.
}
\description{
List containing related DESeqResults objects
}
\note{
Updated 2023-12-18.
}
\author{
Michael Steinbaugh
}
