% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllGenerators.R
\name{DESeqResultsList}
\alias{DESeqResultsList}
\alias{DESeqResultsList,DESeqAnalysis-method}
\alias{DESeqResultsList,DESeqAnalysisList-method}
\alias{DESeqResultsList,SimpleList-method}
\alias{DESeqResultsList,list-method}
\alias{DESeqResultsList,missing-method}
\title{List containing related DESeqResults objects}
\usage{
DESeqResultsList(object, ...)

\S4method{DESeqResultsList}{DESeqAnalysis}(object, lfcShrink = NULL, quiet = FALSE)

\S4method{DESeqResultsList}{DESeqAnalysisList}(object, ...)

\S4method{DESeqResultsList}{SimpleList}(object)

\S4method{DESeqResultsList}{list}(object)

\S4method{DESeqResultsList}{missing}(object)
}
\arguments{
\item{...}{\code{DESeqResults} objects or named \code{list}.}
}
\value{
\code{DESeqResultsList}.
}
\description{
List containing related DESeqResults objects
}
\note{
Updated 2021-03-12.
}
\examples{
data(deseq)
x <- DESeqResultsList(deseq)
print(names(x))
}
