% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,DESeqAnalysis,character,missing-method}
\alias{export,DESeqDataSet,character,missing-method}
\title{Export}
\usage{
export(object, con, format, ...)

\S4method{export}{DESeqAnalysis,character,missing}(
  object,
  con,
  format,
  compress = getOption(x = "acid.export.compress", default = FALSE),
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{DESeqDataSet,character,missing}(
  object,
  con,
  format,
  compress = getOption(x = "acid.export.compress", default = FALSE),
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.}

\item{con}{\code{character(1)}.
Directory path.}

\item{format}{\code{character(1)}, \code{missing}, or \code{NULL}.
\emph{Currently not supported.}}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export
}
\details{
Size-factor normalized coutns and FPKM values are calculated on the fly and
exported automatically.
}
\note{
Updated 2022-09-13.
}
\examples{
data(deseq)

## DESeqAnalysis ====
con <- AcidBase::tempdir2()
x <- export(deseq, con = con)
print(x)
AcidBase::unlink2(con)
}
