% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,DESeqAnalysis-method}
\title{Export}
\usage{
\S4method{export}{DESeqAnalysis}(x, name = NULL, dir = ".",
  compress = FALSE, humanize = FALSE)
}
\arguments{
\item{x}{Object.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{humanize}{\code{logical(1)}.
Make the gene and sample names human readable.
\itemize{
\item Gene names require \code{geneName} column to be defined in
\code{\link[SummarizedExperiment:rowData]{rowData()}}.
\item Sample names require \code{sampleName} column to be defined in
\code{\link[SummarizedExperiment:colData]{colData()}}.
}}
}
\description{
Export data out of R and write to disk.
}
\examples{
data(deseq)

## DESeqAnalysis ====
export(deseq, dir = "example")

## Clean up.
unlink("example", recursive = TRUE)
}
