% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,DESeqDataSet-method}
\alias{export,DESeqAnalysis-method}
\title{Export}
\usage{
\S4method{export}{DESeqDataSet}(x, name = NULL, dir = ".",
  compress = FALSE)

\S4method{export}{DESeqAnalysis}(x, name = NULL, dir = ".",
  compress = FALSE)
}
\arguments{
\item{x}{Object.
An object supporting \code{\link[base:dim]{dim()}}, or a supported class capable
of being coerced to \code{data.frame}, to be written to disk.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}
}
\description{
Export data out of R and write to disk.
}
\examples{
data(deseq)

## DESeqAnalysis ====
export(deseq, dir = "example")

## Clean up.
unlink("example", recursive = TRUE)
}
