% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,DESeqAnalysis,character-method}
\alias{export,DESeqDataSet,character-method}
\title{Export data from R}
\usage{
export(object, con, ...)

\S4method{export}{DESeqAnalysis,character}(object, con, compress = FALSE, overwrite = TRUE, quiet = FALSE)

\S4method{export}{DESeqDataSet,character}(object, con, compress = FALSE, overwrite = TRUE, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{con}{\code{character(1)}.
Directory path.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character} or \code{list} of file paths.
}
\description{
Export data from R
}
\details{
Size-factor normalized coutns and FPKM values are calculated on the fly and
exported automatically.
}
\note{
Updated 2022-09-13.
}
\examples{
data(deseq)

## DESeqAnalysis ====
con <- AcidBase::tempdir2()
x <- export(deseq, con = con)
print(x)
AcidBase::unlink2(con)
}
