% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,DESeqAnalysis-method}
\title{Export}
\usage{
\S4method{export}{DESeqAnalysis}(x, name = NULL, dir = ".",
  compress = FALSE, humanize = FALSE)
}
\arguments{
\item{x}{Object.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{humanize}{\code{logical(1)}.
Automatically convert gene IDs to gene symbols in the row names and sample
IDs to sample names in the column names.}
}
\description{
Export data out of R and write to disk.
}
\details{
This is a wrapper for \code{\link[rio:export]{rio::export()}} that adds support for additional S4
classes in Bioconductor.
}
\examples{
data(deseq)

## DESeqAnalysis ====
export(deseq, dir = "example")
}
