% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{export}
\alias{export}
\alias{export,DESeqDataSet-method}
\alias{export,DESeqAnalysis-method}
\title{Export}
\usage{
export(object, ...)

\S4method{export}{DESeqDataSet}(object, name = NULL, dir = ".", compress = FALSE)

\S4method{export}{DESeqAnalysis}(object, name = NULL, dir = ".", compress = FALSE, lfcShrink = TRUE)
}
\arguments{
\item{object}{Object.}

\item{name}{\code{character(1)}.
Name of the individual effect (coefficient) for building a results table.
Use this argument rather than \code{contrast} for continuous variables.}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}

\item{...}{Additional arguments.}
}
\value{
Invisibily return file paths, either as \code{list} or \code{character}.
}
\description{
Export data out of R and write to disk.
}
\note{
Size-factor normalized coutns and FPKM values are calculated on the fly
and exported automatically.

Updated 2019-11-12.
}
\examples{
data(deseq)

## DESeqAnalysis ====
export(deseq, dir = "example")

## Clean up.
unlink("example", recursive = TRUE)
}
