% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lfcShrink-methods.R
\name{lfcShrink}
\alias{lfcShrink}
\alias{lfcShrink,DESeqDataSet-method}
\title{Return shrunken log2 fold change values?}
\usage{
lfcShrink(object, ...)

\S4method{lfcShrink}{DESeqDataSet}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{logical(1)}.
}
\description{
Return shrunken log2 fold change values?
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)
lfcShrink(deseq)
}
\seealso{
\code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.
}
