% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lfcShrink-methods.R
\name{lfcShrink}
\alias{lfcShrink}
\alias{lfcShrink<-}
\alias{lfcShrink,DESeqAnalysis-method}
\alias{lfcShrink,DESeqAnalysisList-method}
\alias{lfcShrink,DESeqDataSet-method}
\alias{lfcShrink<-,DESeqAnalysis,logical-method}
\alias{lfcShrink<-,DESeqAnalysisList,logical-method}
\title{Shrink log2 fold changes}
\usage{
lfcShrink(object, ...)
lfcShrink(object, ...) <- value

\S4method{lfcShrink}{DESeqAnalysis}(object)

\S4method{lfcShrink}{DESeqAnalysisList}(object)

\S4method{lfcShrink}{DESeqDataSet}(object, ...)

\S4method{lfcShrink}{DESeqAnalysis,logical}(object) <- value

\S4method{lfcShrink}{DESeqAnalysisList,logical}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
Varies, depending on object class.
}
\description{
Shrink log2 fold changes
}
\note{
Updated 2021-06-28.
}
\examples{
data(deseq)

## DESeqAnalysis ====
lfcShrink(deseq)
}
\seealso{
\code{DESeq2::lfcShrink()}.
}
