% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lfcThreshold-methods.R
\name{lfcThreshold}
\alias{lfcThreshold}
\alias{lfcThreshold<-}
\alias{lfcThreshold,DESeqAnalysis-method}
\alias{lfcThreshold,DESeqAnalysisList-method}
\alias{lfcThreshold,DESeqResults-method}
\alias{lfcThreshold,DESeqResultsList-method}
\alias{lfcThreshold<-,DESeqAnalysis,numeric-method}
\alias{lfcThreshold<-,DESeqAnalysis,NULL-method}
\alias{lfcThreshold<-,DESeqAnalysisList,numeric-method}
\alias{lfcThreshold<-,DESeqAnalysisList,NULL-method}
\alias{lfcThreshold<-,DESeqResults,numeric-method}
\alias{lfcThreshold<-,DESeqResultsList,numeric-method}
\title{Log2 fold change threshold}
\usage{
lfcThreshold(object, ...)

lfcThreshold(object, ...) <- value

\S4method{lfcThreshold}{DESeqAnalysis}(object)

\S4method{lfcThreshold}{DESeqAnalysisList}(object)

\S4method{lfcThreshold}{DESeqResults}(object)

\S4method{lfcThreshold}{DESeqResultsList}(object)

\S4method{lfcThreshold}{DESeqAnalysis,numeric}(object) <- value

\S4method{lfcThreshold}{DESeqAnalysis,`NULL`}(object) <- value

\S4method{lfcThreshold}{DESeqAnalysisList,numeric}(object) <- value

\S4method{lfcThreshold}{DESeqAnalysisList,`NULL`}(object) <- value

\S4method{lfcThreshold}{DESeqResults,numeric}(object) <- value

\S4method{lfcThreshold}{DESeqResultsList,numeric}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric(1)}.
}
\description{
Log2 fold change threshold
}
\details{
Assumes \code{0} by default if unset.
}
\note{
Updated 2021-03-12.
}
\examples{
data(deseq)

## DESeqAnalysis ====
lfcThreshold(deseq) <- 0.2
lfcThreshold(deseq)
}
