% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,DESeqAnalysis-method}
\title{Markdown}
\usage{
markdown(object, ...)

\S4method{markdown}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\section{Methods (by class)}{
\itemize{
\item \code{DESeqAnalysis}: List of contrast names.
}}

\examples{
data(deseq)

# DESeqAnalysis ====
markdown(deseq)
}
