% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,DESeqAnalysis-method}
\alias{markdown,DESeqResultsTables-method}
\title{Markdown}
\usage{
\S4method{markdown}{DESeqAnalysis}(object)

\S4method{markdown}{DESeqResultsTables}(object, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). Markdown header level.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\section{Methods (by class)}{
\itemize{
\item \code{DESeqAnalysis}: List of contrast names.

\item \code{DESeqResultsTables}: File paths (if exported) and top tables.
}}

\examples{
data(deseq)

# DESeqAnalysis ====
markdown(deseq)

## DESeqResultsTables ====
x <- DESeqResultsTables(deseq)
markdown(x)
}
