% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMetadataToFiles.R
\name{matchMetadataToFiles}
\alias{matchMetadataToFiles}
\title{Match user metadata to file names used for tximport}
\usage{
matchMetadataToFiles(metadata, files)
}
\arguments{
\item{metadata}{\code{data.frame}.
User-defined metadata. The function assumes that sample identifiers are
defined in the first metadata column.}

\item{files}{\code{character}.
Quant file paths passed to \code{\link[tximport:tximport]{tximport()}}.}
}
\value{
\code{data.frame}.
Contains 1:1 file to metadata sample identifier mappings.
}
\description{
Match user metadata to file names used for tximport
}
\note{
Updated 2019-10-09.
}
\examples{
metadata <- data.frame(
    sampleID = paste(seq_len(4L), "sample", LETTERS[seq_len(4L)], sep = "_"),
    condition = rep(LETTERS[seq_len(2L)], times = 2L)
)
## The function will match regardless of row order in user metadata.
metadata <- metadata[nrow(metadata):1L, ]
files <- file.path(
    "salmon",
    paste(seq_len(4L), "sample", LETTERS[seq_len(4L)], sep = "-"),
    "quant.sf"
)
print(metadata)
print(files)
matchMetadataToFiles(metadata = metadata, files = files)
}
