% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
