% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}

\item{object}{Object.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or positive of \code{DESeqResults}.}

\item{...}{Additional arguments}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
