% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{contrastSamples}{\code{logical(1)}. \strong{Experimental}. Only include the samples
used to define the contrast passed to \code{DESeq2::results}. This setting
will break for complex DESeq2 contrasts (e.g. interaction effect).}

\item{lfcShrink}{\code{logical(1)}. Use shrunken log2 fold change (LFC) values.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
