% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{DESeqDataSet}{\code{DESeqDataSet} or \code{NULL}.}

\item{DESeqTransform}{\code{DESeqTransform}.}

\item{contrast}{\code{character}.
A character vector with exactly 3 elements:
\enumerate{
\item Name of factor in the design formula.
\item Name of numerator level for the fold change.
\item Name of denominator level for the fold change.
}

See \code{\link[DESeq2:results]{DESeq2::results()}} for details.}

\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}

\item{i}{\code{character(1)} or \code{integer(1)}.
\code{DESeqResults} contrast name or position in \code{results} slot.}

\item{lfcShrink}{\code{logical(1)} or \code{NULL}.
Use shrunken log2 fold change (LFC) values.
If \code{NULL}, inherits value defined in \code{\link[=lfcShrink]{lfcShrink()}}.}

\item{limits}{\code{list(2)}.
Named list containing \code{"x"} and \code{"y"} that define the lower and upper
limits for each axis. Set automatically by default when left \code{NULL}.}

\item{name}{\code{character(1)}.
Name of the individual effect (coefficient) for building a results table.
Use this argument rather than \code{contrast} for continuous variables.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
