% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotBaseMean-methods.R
\docType{methods}
\name{plotBaseMean}
\alias{plotBaseMean}
\alias{plotBaseMean,DESeqDataSet-method}
\alias{plotBaseMean,DESeqResults-method}
\alias{plotBaseMean,DESeqAnalysis-method}
\title{Plot base mean distribution}
\usage{
plotBaseMean(object, ...)

\S4method{plotBaseMean}{DESeqDataSet}(object, nonzero = TRUE,
  trans = c("log10", "log2", "identity"))

\S4method{plotBaseMean}{DESeqResults}(object, nonzero = TRUE,
  trans = c("log10", "log2", "identity"))

\S4method{plotBaseMean}{DESeqAnalysis}(object, nonzero = TRUE,
  trans = c("log10", "log2", "identity"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{nonzero}{\code{logical(1)}.
Remove zero-count features (genes).}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}
}
\value{
\code{ggplot}.
}
\description{
The base mean is the mean of normalized counts of all samples, normalizing
for sequencing depth.
}
\note{
Updated 2019-09-10.
}
\examples{
data(deseq)
plotBaseMean(deseq)
}
\seealso{
\itemize{
\item https://support.bioconductor.org/p/75244/
}
}
