% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotBaseMean-methods.R
\name{plotBaseMean}
\alias{plotBaseMean}
\alias{plotBaseMean,DESeqAnalysis-method}
\alias{plotBaseMean,DESeqDataSet-method}
\alias{plotBaseMean,DESeqResults-method}
\title{Plot base mean distribution}
\usage{
plotBaseMean(object, ...)

\S4method{plotBaseMean}{DESeqAnalysis}(
  object,
  nonzero = TRUE,
  trans = c("log10", "log2", "identity"),
  summary = TRUE,
  labels = list(title = "Base mean distribution", subtitle = NULL)
)

\S4method{plotBaseMean}{DESeqDataSet}(
  object,
  nonzero = TRUE,
  trans = c("log10", "log2", "identity"),
  summary = TRUE,
  labels = list(title = "Base mean distribution", subtitle = NULL)
)

\S4method{plotBaseMean}{DESeqResults}(
  object,
  nonzero = TRUE,
  trans = c("log10", "log2", "identity"),
  summary = TRUE,
  labels = list(title = "Base mean distribution", subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{nonzero}{\code{logical(1)}.
Remove zero-count features (genes).}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{summary}{\code{logical(1)}.
Include distribution summary statistics as lines on the plot.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}
}
\value{
\code{ggplot}.
}
\description{
The base mean is the mean of normalized counts of all samples, normalizing
for sequencing depth.
}
\section{Functions}{
\itemize{
\item \code{plotBaseMean,DESeqAnalysis-method}: Passes to \code{DESeqDataSet} method.

\item \code{plotBaseMean,DESeqDataSet-method}: Generates row means of normalized counts.
This value corresponds to the \code{baseMean} column of \code{DESeqResults}.

\item \code{plotBaseMean,DESeqResults-method}: Uses \code{baseMean} column of results.
}}

\note{
Updated 2022-05-17.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotBaseMean(deseq)
}
\seealso{
\itemize{
\item https://support.bioconductor.org/p/75244/
\item \code{summary()}.
}
}
