% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotBaseMean-methods.R
\docType{methods}
\name{plotBaseMean}
\alias{plotBaseMean}
\alias{plotBaseMean,DESeqDataSet-method}
\alias{plotBaseMean,DESeqResults-method}
\alias{plotBaseMean,DESeqAnalysis-method}
\title{Plot base mean distribution}
\usage{
plotBaseMean(object, ...)

\S4method{plotBaseMean}{DESeqDataSet}(object)

\S4method{plotBaseMean}{DESeqResults}(object)

\S4method{plotBaseMean}{DESeqAnalysis}(object)
}
\value{
\code{ggplot}.
}
\description{
The base mean is the mean of normalized counts of all samples, normalizing
for sequencing depth.
}
\note{
Updated 2019-07-30.
}
\examples{
data(deseq)
plotBaseMean(deseq)
}
\seealso{
\itemize{
\item https://support.bioconductor.org/p/75244/
}
}
