% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotBaseMean-methods.R
\name{plotBaseMean}
\alias{plotBaseMean}
\alias{plotBaseMean,DESeqAnalysis-method}
\alias{plotBaseMean,DESeqDataSet-method}
\alias{plotBaseMean,DESeqResults-method}
\alias{plotBaseMean,numeric-method}
\title{Plot base mean distribution}
\usage{
plotBaseMean(object, ...)

\S4method{plotBaseMean}{DESeqAnalysis}(object, ...)

\S4method{plotBaseMean}{DESeqDataSet}(object, ...)

\S4method{plotBaseMean}{DESeqResults}(object, ...)

\S4method{plotBaseMean}{numeric}(
  object,
  nonzero = TRUE,
  trans = c("log10", "log2", "identity"),
  summary = TRUE,
  color = getOption(x = "acid.color.discrete", default = getOption(x =
    "ggplot2.discrete.colour", default = AcidPlots::scale_color_synesthesia_d())),
  labels = list(title = "Base mean distribution", subtitle = NULL, x =
    "average expression across all samples", y = "density", color = "summary")
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{nonzero}{\code{logical(1)}.
Remove zero-count features (genes).}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{summary}{\code{logical(1)}.
Include distribution summary statistics as lines on the plot.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}
}
\value{
\code{ggplot}.
}
\description{
The base mean is the mean of normalized counts of all samples, normalizing
for sequencing depth.
}
\section{Functions}{
\itemize{
\item \code{plotBaseMean,DESeqAnalysis-method}: Passes to \code{DESeqDataSet} method.

\item \code{plotBaseMean,DESeqDataSet-method}: Generates row means of normalized counts.
This value corresponds to the \code{baseMean} column of \code{DESeqResults}.
Passes to \code{numeric} method.

\item \code{plotBaseMean,DESeqResults-method}: Uses \code{baseMean} column of results.
Passes to \code{numeric} method.
}}

\note{
Updated 2021-03-15.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotBaseMean(deseq)
}
\seealso{
\itemize{
\item https://support.bioconductor.org/p/75244/
\item \code{\link[base:summary]{summary()}}.
}
}
