% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotContrast-methods.R
\name{plotContrast}
\alias{plotContrast}
\alias{plotContrast,DESeqAnalysis-method}
\title{Plot differential expression contrast}
\usage{
plotContrast(object, ...)

\S4method{plotContrast}{DESeqAnalysis}(object, i)
}
\description{
Plot differential expression contrast
}
\note{
Updated 2021-08-02.
}
\examples{
data(deseq)
plotContrast(deseq, i = 1L)

}
\seealso{

}
