% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotContrastScatter-methods.R
\name{plotContrastScatter}
\alias{plotContrastScatter}
\alias{plotContrastScatter,DESeqAnalysis-method}
\title{Plot scatterplot of differential expression contrast}
\usage{
plotContrastScatter(object, ...)

\S4method{plotContrastScatter}{DESeqAnalysis}(
  object,
  i,
  direction = c("both", "up", "down"),
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  pointColor = c(downregulated = AcidPlots::lightPalette[["purple"]], upregulated =
    AcidPlots::lightPalette[["orange"]], nonsignificant =
    AcidPlots::lightPalette[["gray"]]),
  pointSize = 2L,
  pointAlpha = 0.8,
  trans = c("log2", "log10", "identity"),
  limits = list(x = NULL, y = NULL),
  labels = list(title = TRUE, subtitle = NULL, x = TRUE, y = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{limits}{\code{list(2)}.
Named list containing \code{"x"} and \code{"y"} that define the lower and upper
limits for each axis. Set automatically by default when left \code{NULL}.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot scatterplot of differential expression contrast
}
\note{
Updated 2022-04-15.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotContrastScatter(deseq, i = 1L)
}
\seealso{
\itemize{
\item https://doi.org/10.1084/jem.20200829
}
}
