% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts,DESeqDataSet-method}
\alias{plotCounts,DESeqAnalysis-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

\S4method{plotCounts}{DESeqDataSet}(object, ...)

\S4method{plotCounts}{DESeqAnalysis}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\section{Functions}{
\itemize{
\item \code{plotCounts,DESeqDataSet-method}: Automatically plots size factor (i.e. library size)
normalized counts. Arguments pass through to \code{SummarizedExperiment} method
defined in acidplots package.

\item \code{plotCounts,DESeqAnalysis-method}: Passes to \code{DESeqDataSet} method.
}}

\note{
Updated 2019-09-17.
}
\examples{
data(deseq)

## Get genes from working example.
res <- results(deseq, i = 1L)
genes <- head(rownames(res))
print(genes)

## DESeqAnalysis ====
plotCounts(deseq, genes = genes, style = "facet")
plotCounts(deseq, genes = genes, style = "wide")
}
\seealso{
\code{DESeq2::plotCounts()}.
}
