% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGHeatmap-methods.R
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults-method}
\alias{plotDEGHeatmap,DESeqAnalysis-method}
\title{Differentially expressed gene heatmap}
\usage{
plotDEGHeatmap(object, ...)

\S4method{plotDEGHeatmap}{DESeqResults}(
  object,
  DESeqTransform,
  direction = c("both", "up", "down"),
  ...
)

\S4method{plotDEGHeatmap}{DESeqAnalysis}(object, i, contrastSamples = FALSE, lfcShrink = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{DESeqTransform}{\code{DESeqTransform}.}

\item{direction}{\code{character(1)}.
Plot \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}
}
\value{
Plot.
}
\description{
This function is an extension of \code{\link[=plotHeatmap]{plotHeatmap()}} that is optimized for
automatic handling differentially expressed genes, rather than requiring
manual input of a gene vector. All of the optional parameters for
\code{\link[=plotHeatmap]{plotHeatmap()}} are also supported by this function.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:colData]{colData()}} of the \code{counts} argument, which
must contain a \code{SummarizedExperiment}.
}
\section{Functions}{
\itemize{
\item \code{plotDEGHeatmap,DESeqResults-method}: Passes to \code{plotHeatmap()} \code{SummarizedExperiment}
method defined in acidplots.

\item \code{plotDEGHeatmap,DESeqAnalysis-method}: Passes to \code{DESeqResults} method.
}}

\note{
Updated 2019-11-19.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGHeatmap(deseq, i = 1L)
}
