% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGHeatmap-methods.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqAnalysis-method}
\title{Differentially expressed gene heatmap}
\usage{
\S4method{plotDEGHeatmap}{DESeqAnalysis}(object, results,
  contrastSamples = FALSE, direction = c("both", "up", "down"),
  scale = c("row", "column", "none"), clusteringMethod = "ward.D2",
  clusterRows = TRUE, clusterCols = TRUE, interestingGroups = NULL,
  showRownames = FALSE, showColnames = TRUE, treeheightRow = 50L,
  treeheightCol = 50L, color = viridis::viridis,
  legendColor = viridis::viridis, borderColor = NULL, title = NULL,
  ...)
}
\arguments{
\item{object}{Object.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or positive of \code{DESeqResults}.}

\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{scale}{\code{character(1)}.
Whether the values should be centered and scaled in either the row or
column direction ("\code{row}", "\code{column}"), or remain unscaled ("\code{none}").}

\item{clusteringMethod}{\code{character(1)}.
Clustering method. Accepts the same values as \code{stats::hclust}.}

\item{clusterRows}{\code{logical(1)}.
Arrange with hierarchical clustering.}

\item{clusterCols}{\code{logical(1)}.
Arrange with hierarchical clustering.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{showRownames}{\code{logical(1)}.
Show row or column names.}

\item{showColnames}{\code{logical(1)}.
Show row or column names.}

\item{treeheightRow}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{color}{\code{function}, \code{character}, or \code{NULL}.
Hexadecimal color function or values to use for plot.

We generally recommend these hexadecimal functions from the viridis
package:
\itemize{
\item \code{viridis::viridis()} (\emph{default}).
\item \code{viridis::inferno()}.
\item \code{viridis::magma()}.
\item \code{viridis::plasma()}.
}

Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.}

\item{legendColor}{\code{function} or \code{NULL}.
Hexadecimal color function to use for legend labels. Note that hexadecimal
values are not supported. If set \code{NULL}, will use the default pheatmap
colors.}

\item{borderColor}{\code{character(1)} or \code{NULL}.
Border color. Disabled by default for improved aesthetics.}

\item{title}{\code{character(1)} or \code{NULL}.
Plot title.}

\item{...}{Additional arguments.}
}
\description{
This function is an extension of \code{\link[=plotHeatmap]{plotHeatmap()}} that is optimized for
automatic handling differentially expressed genes, rather than requiring
manual input of a gene vector. All of the optional parameters for
\code{\link[=plotHeatmap]{plotHeatmap()}} are also supported by this function.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:colData]{colData()}} of the \code{counts} argument, which
must contain a \code{SummarizedExperiment}.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGHeatmap(deseq, results = 1L)
}
