% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGHeatmap-methods.R
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults-method}
\alias{plotDEGHeatmap,DESeqAnalysis-method}
\title{Differentially expressed gene heatmap}
\usage{
plotDEGHeatmap(object, ...)

\S4method{plotDEGHeatmap}{DESeqResults}(
  object,
  DESeqTransform,
  alphaThreshold = NULL,
  lfcThreshold = NULL,
  baseMeanThreshold = NULL,
  direction = c("both", "up", "down"),
  title = TRUE,
  subtitle = TRUE,
  ...
)

\S4method{plotDEGHeatmap}{DESeqAnalysis}(
  object,
  i,
  contrastSamples = FALSE,
  alphaThreshold = NULL,
  lfcShrink = NULL,
  lfcThreshold = NULL,
  baseMeanThreshold = NULL,
  ...
)
}
\arguments{
\item{object}{Object.}

\item{DESeqTransform}{\code{DESeqTransform}.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{title}{\code{logical(1)}, \code{character(1)}.
Include contrast name as title?
Can manually define as \code{character}.}

\item{subtitle}{\code{logical(1)}.
Include subtitle containing DEG information?}

\item{...}{Additional arguments.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}

\item{lfcShrink}{\code{logical(1)} or \code{NULL}.
Use shrunken log2 fold change (LFC) values.
If \code{NULL}, inherits value defined in \code{\link[=lfcShrink]{lfcShrink()}}.}
}
\value{
Plot.
}
\description{
This function is an extension of \code{\link[acidgenerics:plotHeatmap]{plotHeatmap()}} that is optimized for
automatic handling differentially expressed genes, rather than requiring
manual input of a gene vector. All of the optional parameters for
\code{\link[acidgenerics:plotHeatmap]{plotHeatmap()}} are also supported by this function.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} of the \code{counts} argument, which
must contain a \code{SummarizedExperiment}.
}
\section{Functions}{
\itemize{
\item \code{plotDEGHeatmap,DESeqResults-method}: Passes to \code{plotHeatmap()} \code{SummarizedExperiment}
method defined in acidplots.

\item \code{plotDEGHeatmap,DESeqAnalysis-method}: Passes to \code{DESeqResults} method.
}}

\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGHeatmap(deseq, i = 1L)
}
