% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA-methods.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults-method}
\alias{plotDEGPCA,DESeqAnalysis-method}
\title{Plot DEG PCA}
\usage{
\S4method{plotDEGPCA}{DESeqAnalysis}(object, results, counts,
  direction = c("both", "up", "down"), interestingGroups = NULL,
  label = getOption("basejump.label", FALSE),
  pointColor = getOption("basejump.color.discrete", NULL),
  pointSize = getOption("basejump.point.size", 4L),
  return = c("ggplot", "DataFrame"), ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{\code{DESeqTransform}.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{pointColor}{\code{string}. Default point color for the plot.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{return}{\code{string}. Return type. Uses \code{base::match.arg()}
internally and defaults to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot DEG PCA
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGPCA(deseq)

## DESeqResults ====
plotDEGPCA(
    object = as(deseq, "DESeqResults"),
    counts = as(deseq, "DESeqTransform")
)
}
