% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA-methods.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults-method}
\alias{plotDEGPCA,DESeqAnalysis-method}
\title{Plot differentially expressed gene principal component analysis}
\usage{
plotDEGPCA(object, ...)

\S4method{plotDEGPCA}{DESeqResults}(object, DESeqTransform,
  direction = c("both", "up", "down"), object, assay = 1L,
  interestingGroups = NULL, label = getOption(x = "acid.label", default
  = FALSE), color = getOption(x = "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), pointSize = getOption(x =
  "acid.point.size", default = 3L), return = c("ggplot", "DataFrame"),
  ...)

\S4method{plotDEGPCA}{DESeqAnalysis}(object, results,
  contrastSamples = FALSE, direction = c("both", "up", "down"),
  assay = 1L, interestingGroups = NULL, label = getOption(x =
  "acid.label", default = FALSE), color = getOption(x =
  "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), pointSize = getOption(x =
  "acid.point.size", default = 3L), return = c("ggplot", "DataFrame"),
  ...)
}
\arguments{
\item{object}{Object.}

\item{DESeqTransform}{\code{DESeqTransform}.}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{color}{\code{function}, \code{character}, or \code{NULL}.
Hexadecimal color function or values to use for plot.

We generally recommend these hexadecimal functions from the viridis
package, in addition to our \code{\link[=synesthesia]{synesthesia()}} palette:
\itemize{
\item \code{viridis::viridis()}.
\item \code{viridis::inferno()}.
\item \code{viridis::magma()}.
\item \code{viridis::plasma()}.
}

Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}

\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}
}
\value{
Plot.
}
\description{
This function is an extension of \code{\link[BiocGenerics:plotPCA]{plotPCA()}} that is
optimized for automatic handling of differentially expressed genes, rather
than requiring manual input of a gene vector or subset object.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:colData]{colData()}} of the \code{counts} argument, which
must contain/extend a \code{SummarizedExperiment}.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGPCA(deseq, results = 1L)
}
