% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA-methods.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults-method}
\alias{plotDEGPCA,DESeqAnalysis-method}
\title{Plot DEG PCA}
\usage{
\S4method{plotDEGPCA}{DESeqResults}(object, counts, direction = c("both",
  "up", "down"), object, assay = 1L, interestingGroups = NULL,
  label = getOption("basejump.label", FALSE),
  color = getOption("basejump.color.discrete", NULL),
  pointSize = getOption("basejump.point.size", 3L),
  return = c("ggplot", "DataFrame"), ...)

\S4method{plotDEGPCA}{DESeqAnalysis}(object, results, counts,
  direction = c("both", "up", "down"), assay = 1L,
  interestingGroups = NULL, label = getOption("basejump.label", FALSE),
  color = getOption("basejump.color.discrete", NULL),
  pointSize = getOption("basejump.point.size", 3L),
  return = c("ggplot", "DataFrame"), ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{\code{DESeqTransform}.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{assay}{\code{scalar}. Name or index of count matrix slotted in
\code{SummarizedExperiment::assays()}. When passing in a \code{string}, the name must
be defined in \code{SummarizedExperiment::assayNames()}.}

\item{interestingGroups}{\code{character}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Point size for dots in the plot.}

\item{return}{\code{string}. Return type. Uses \code{base::match.arg()}
internally and defaults to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
\code{ggplot}.
}
\description{
Plot DEG PCA
}
\examples{
data(deseq)

## DESeqAnalysis (recommended) ====
plotDEGPCA(deseq)

## DESeqResults (legacy; not recommended) ====
plotDEGPCA(
    object = as(deseq, "DESeqResults"),
    counts = as(deseq, "DESeqTransform")
)
}
