% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA-methods.R
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults-method}
\alias{plotDEGPCA,DESeqAnalysis-method}
\title{Plot differentially expressed gene principal component analysis}
\usage{
plotDEGPCA(object, ...)

\S4method{plotDEGPCA}{DESeqResults}(
  object,
  DESeqTransform,
  alpha = NULL,
  lfcThreshold = NULL,
  direction = c("both", "up", "down"),
  ...
)

\S4method{plotDEGPCA}{DESeqAnalysis}(object, i, contrastSamples = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{DESeqTransform}{\code{DESeqTransform}.}

\item{alpha}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}
}
\value{
Plot.
}
\description{
This function is an extension of \code{\link[BiocGenerics:plotPCA]{plotPCA()}} that is
optimized for automatic handling of differentially expressed genes, rather
than requiring manual input of a gene vector or subset object.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} of the \code{counts} argument, which
must contain/extend a \code{SummarizedExperiment}.
}
\section{Functions}{
\itemize{
\item \code{plotDEGPCA,DESeqAnalysis-method}: Passes to \code{DESeqResults} method.
}}

\note{
Updated 2019-12-18.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGPCA(deseq, i = 1L)
}
