% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDegPca-methods.R
\name{plotDegPca}
\alias{plotDegPca}
\alias{plotDegPca,DESeqAnalysis-method}
\alias{plotDegPca,DESeqResults-method}
\title{Plot differentially expressed gene principal component analysis}
\usage{
plotDegPca(object, ...)

\S4method{plotDegPca}{DESeqAnalysis}(
  object,
  i,
  contrastSamples = FALSE,
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  ...
)

\S4method{plotDegPca}{DESeqResults}(
  object,
  DESeqTransform,
  direction = c("both", "up", "down"),
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  ...
)
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{contrastSamples}{\code{logical(1)}.
Only include the samples used to define the contrast passed to
\code{DESeq2::results()}. This setting will break for complex DESeq2 contrasts
(e.g. interaction effect).}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{...}{Additional arguments.}

\item{DESeqTransform}{\code{DESeqTransform}.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}
}
\value{
Plot.
}
\description{
This function is an extension of \code{plotPCA()} that is optimized for automatic
handling of differentially expressed genes, rather than requiring manual
input of a gene vector or subset object.
}
\details{
To adjust the annotation columns, modify the \code{colData()} of the \code{counts}
argument, which must contain/extend a \code{SummarizedExperiment}.
}
\section{Functions}{
\itemize{
\item \code{plotDegPca(DESeqAnalysis)}: Passes to \code{DESeqResults} method.

}}
\note{
Updated 2021-08-02.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDegPca(deseq, i = 1L)
}
