% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA-methods.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults-method}
\alias{plotDEGPCA,DESeqAnalysis-method}
\title{Plot differentially expressed gene principal component analysis}
\usage{
\S4method{plotDEGPCA}{DESeqResults}(object, counts, direction = c("both",
  "up", "down"), object, assay = 1L, interestingGroups = NULL,
  label = getOption("basejump.label", default = FALSE),
  color = getOption("basejump.color.discrete", default = NULL),
  pointSize = getOption("basejump.point.size", default = 3L),
  return = c("ggplot", "DataFrame"), ...)

\S4method{plotDEGPCA}{DESeqAnalysis}(object, results,
  contrastSamples = FALSE, direction = c("both", "up", "down"),
  assay = 1L, interestingGroups = NULL,
  label = getOption("basejump.label", default = FALSE),
  color = getOption("basejump.color.discrete", default = NULL),
  pointSize = getOption("basejump.point.size", default = 3L),
  return = c("ggplot", "DataFrame"), ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{\code{DESeqTransform}.
Variance-stabilized counts suitable for heatmap.
Object rownames must be identical to corresponding \code{DESeqResults}.}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{assay}{\code{vector(1)}.
Name or index of count matrix slotted in
\code{\link[SummarizedExperiment:assays]{assays()}}. When passing in a string, the
name must be defined in
\code{\link[SummarizedExperiment:assayNames]{assayNames()}}.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{color}{\code{function}, \code{character}, or \code{NULL}.
Hexadecimal color function or values to use for plot.

We generally recommend these hexadecimal functions from the viridis
package:
\itemize{
\item \code{viridis::viridis()} (\emph{default}).
\item \code{viridis::inferno()}.
\item \code{viridis::magma()}.
\item \code{viridis::plasma()}.
}

Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}

\item{contrastSamples}{\code{logical(1)}.
\strong{Experimental}. Only include the samples used to define the contrast
passed to \code{\link[DESeq2:results]{DESeq2::results()}}. This setting will break for complex DESeq2
contrasts (e.g. interaction effect).}
}
\description{
This function is an extension of \code{\link[BiocGenerics:plotPCA]{plotPCA()}} that is
optimized for automatic handling of differentially expressed genes, rather
than requiring manual input of a gene vector or subset object.
}
\details{
To adjust the annotation columns, modify the
\code{\link[SummarizedExperiment:colData]{colData()}} of the \code{counts} argument, which
must contain/extend a \code{SummarizedExperiment}.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGPCA(deseq, results = 1L)
}
