% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDEGStackedBar-methods.R
\name{plotDEGStackedBar}
\alias{plotDEGStackedBar}
\alias{plotDEGStackedBar,DESeqAnalysis-method}
\title{Stacked bar plot of DEGs}
\usage{
plotDEGStackedBar(object, ...)

\S4method{plotDEGStackedBar}{DESeqAnalysis}(
  object,
  alpha = NULL,
  lfcThreshold = NULL,
  fill = getOption(x = "acid.fill.discrete", default =
    acidplots::scale_fill_synesthesia_d()),
  flip = TRUE
)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log fold change ratio (base 2) cutoff threshold. If left \code{NULL}, will use
the cutoff defined in the object.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}
}
\description{
Stacked bar plot of DEGs
}
\note{
Updated 2019-12-13.
}
