% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDegStackedBar-methods.R
\name{plotDegStackedBar}
\alias{plotDegStackedBar}
\alias{plotDegStackedBar,DESeqAnalysis-method}
\title{Stacked bar plot of DEGs}
\usage{
plotDegStackedBar(object, ...)

\S4method{plotDegStackedBar}{DESeqAnalysis}(
  object,
  i = NULL,
  direction = c("both", "up", "down"),
  orderBySize = FALSE,
  label = TRUE,
  flip = TRUE
)
}
\arguments{
\item{object}{Object.}

\item{i}{\code{character}, \code{numeric}, or \code{NULL}.
Names or range of results.
If set \code{NULL}, include all results.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{orderBySize}{\code{logical(1)}.
Order contrasts by DEG set size.}

\item{label}{\code{logical(1)}.
Label the number of DEGs per contrast on the plot.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Stacked bar plot of DEGs
}
\note{
Updated 2022-05-18.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDegStackedBar(deseq)
}
