% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDEGUpset-methods.R
\docType{methods}
\name{plotDEGUpset}
\alias{plotDEGUpset}
\alias{plotDEGUpset,DESeqAnalysis-method}
\title{UpSet plot of directional DEG intersections across contrasts}
\usage{
plotDEGUpset(object, ...)

\S4method{plotDEGUpset}{DESeqAnalysis}(object)
}
\arguments{
\item{...}{Additional arguments.}
}
\description{
UpSet plot of directional DEG intersections across contrasts
}
\examples{
data(deseq)
plotDEGUpset(deseq)
}
