% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGUpset-methods.R
\name{plotDEGUpset}
\alias{plotDEGUpset}
\alias{plotDEGUpset,DESeqAnalysis-method}
\title{UpSet plot of directional DEG intersections across contrasts}
\usage{
plotDEGUpset(object, ...)

\S4method{plotDEGUpset}{DESeqAnalysis}(object, direction = c("both", "up", "down"))
}
\arguments{
\item{object}{Object.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional DEG intersections across contrasts
}
\note{
Updated 2020-08-05.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGUpset(deseq)
}
