% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGUpset-methods.R
\name{plotDEGUpset}
\alias{plotDEGUpset}
\alias{plotDEGUpset,DESeqAnalysis-method}
\title{UpSet plot of directional DEG intersections across contrasts}
\usage{
plotDEGUpset(object, ...)

\S4method{plotDEGUpset}{DESeqAnalysis}(
  object,
  alphaThreshold = NULL,
  lfcThreshold = NULL,
  baseMeanThreshold = NULL,
  direction = c("both", "up", "down")
)
}
\arguments{
\item{object}{Object.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional DEG intersections across contrasts
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)
plotDEGUpset(deseq)
}
