% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGUpset-methods.R
\name{plotDEGUpset}
\alias{plotDEGUpset}
\alias{plotDEGUpset,DESeqAnalysis-method}
\title{UpSet plot of directional DEG intersections across contrasts}
\usage{
plotDEGUpset(object, ...)

\S4method{plotDEGUpset}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Plot.
}
\description{
UpSet plot of directional DEG intersections across contrasts
}
\note{
Updated 2019-11-19.
}
\examples{
data(deseq)
plotDEGUpset(deseq)
}
