% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotLFC-methods.R
\name{plotLFC}
\alias{plotLFC}
\alias{plotLFC,DESeqAnalysis-method}
\title{Plot log2 fold change distributions}
\usage{
plotLFC(object, ...)

\S4method{plotLFC}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot log2 fold change distributions
}
\note{
Updated 2022-03-30.

Plot the log2 fold change distribution per contrast.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotLFC(deseq)

## Plot expression of specific genes.
genes <- head(rownames(as.DESeqDataSet(deseq)))
plotLFC(deseq, genes = genes)
}
