% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMa-methods.R
\name{plotMa}
\alias{plotMa}
\alias{plotMa,DESeqAnalysis-method}
\alias{plotMa,DESeqResults-method}
\title{MA plot}
\usage{
plotMa(object, ...)

\S4method{plotMa}{DESeqAnalysis}(
  object,
  i,
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  genes = NULL,
  ntop = 0L,
  ...
)

\S4method{plotMa}{DESeqResults}(
  object,
  direction = c("both", "up", "down"),
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  genes = NULL,
  ntop = 0L,
  pointColor = c(downregulated = AcidPlots::lightPalette[["purple"]], upregulated =
    AcidPlots::lightPalette[["orange"]], nonsignificant =
    AcidPlots::lightPalette[["gray"]]),
  pointSize = 2L,
  pointAlpha = 0.8,
  limits = list(x = NULL, y = NULL),
  labels = list(title = TRUE, subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{ntop}{\code{integer(1)}.
Number of top features to label.}

\item{...}{Additional arguments.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{limits}{\code{list(2)}.
Named list containing \code{"x"} and \code{"y"} that define the lower and upper
limits for each axis. Set automatically by default when left \code{NULL}.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}
}
\value{
\code{ggplot}.
}
\description{
The plot visualizes the differences between measurements taken
in two samples, by transforming the data onto \strong{M} (log ratio) and \strong{A}
(mean average) scales.
}
\details{
An MA plot is an application of a Bland–Altman plot for visual
representation of genomic data. The plot visualizes the differences between
measurements taken in two samples, by transforming the data onto
M (log ratio) and A (mean average) scales, then plotting these values.
}
\section{Functions}{
\itemize{
\item \code{plotMa(DESeqAnalysis)}: Passes to \code{DESeqResults} method, with \code{gene2symbol}
argument automatically defined.

}}
\note{
Updated 2023-12-18.
}
\examples{
data(deseq)

## Get genes from DESeqDataSet.
dds <- as(deseq, "DESeqDataSet")
genes <- head(rownames(dds))
print(genes)

## DESeqAnalysis ====
plotMa(deseq, i = 1L)

## Customize the colors.
plotMa(
    object = deseq,
    i = 1L,
    pointColor = c(
        downregulated = "red",
        nonsignificant = "black",
        upregulated = "green"
    )
)

## Directional support (up or down).
plotMa(deseq, i = 1L, direction = "up", ntop = 5L)
plotMa(deseq, i = 1L, direction = "down", ntop = 5L)

## Label genes manually.
## Note that either gene IDs or names (symbols) are supported.
plotMa(deseq, i = 1L, genes = genes)
}
\seealso{
\code{DESeq2::plotMA()}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
