% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMA-methods.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA2}
\alias{plotMA,DESeqResults-method}
\alias{plotMA,DESeqAnalysis-method}
\alias{plotMA2,DESeqDataSet-method}
\alias{plotMA2,DESeqResults-method}
\alias{plotMeanAverage}
\title{MA-plot: plot differences versus averages for high-throughput data}
\usage{
plotMA(object, ...)

plotMA2(object, ...)

\S4method{plotMA}{DESeqResults}(object, genes = NULL,
  gene2symbol = NULL, ntop = 0L, direction = c("both", "up", "down"),
  pointColor = c(downregulated = "darkorchid3", upregulated =
  "darkorange2", nonsignificant = "gray50"), pointSize = 2L,
  pointAlpha = 0.7, return = c("ggplot", "DataFrame"))

\S4method{plotMA}{DESeqAnalysis}(object, results, lfcShrink = TRUE,
  genes = NULL, ntop = 0L, direction = c("both", "up", "down"),
  pointColor = c(downregulated = "darkorchid3", upregulated =
  "darkorange2", nonsignificant = "gray50"), pointSize = 2L,
  pointAlpha = 0.7, return = c("ggplot", "DataFrame"))

\S4method{plotMA2}{DESeqDataSet}(object, alpha = 0.1, main = "",
  xlab = "mean of normalized counts", ylim, MLE = FALSE, ...)

\S4method{plotMA2}{DESeqResults}(object, alpha, main = "",
  xlab = "mean of normalized counts", ylim, MLE = FALSE, ...)
}
\arguments{
\item{object}{
    A data object, typically containing count values from an RNA-Seq experiment or microarray intensity values.
  }

\item{...}{Additional arguments.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{ntop}{\code{integer(1)}.
Number of top genes to label.}

\item{direction}{\code{character(1)}.
Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size. In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)}.
Alpha transparency level. Must be a proportion (0-1).}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}

\item{alpha}{the significance level for thresholding adjusted p-values}

\item{main}{optional title for the plot}

\item{xlab}{optional defaults to "mean of normalized counts"}

\item{ylim}{optional y limits}

\item{MLE}{if \code{betaPrior=TRUE} was used,
whether to plot the MLE (unshrunken estimates), defaults to FALSE.
Requires that \code{\link{results}} was run with \code{addMLE=TRUE}.
Note that the MLE will be plotted regardless of this argument,
if DESeq() was run with \code{betaPrior=FALSE}. See \code{\link{lfcShrink}}
for examples on how to plot shrunken log2 fold changes.}
}
\value{
\code{ggplot}.
}
\description{
A generic function which produces an MA-plot for an object containing microarray, RNA-Seq
  or other data.
}
\details{
An MA plot is an application of a Bland–Altman plot for visual representation
of genomic data. The plot visualizes the differences between measurements
taken in two samples, by transforming the data onto M (log ratio) and A
(mean average) scales, then plotting these values.
}
\note{
We are not allowing post hoc \code{alpha} or \code{lfcThreshold} cutoffs here.
}
\section{plotMA2 aliases}{


Aliased methods for original \code{\link[DESeq2:plotMA]{DESeq2::plotMA()}} S4 methods, which us
geneplotter instead of ggplot2. I prefer using ggplot2 instead, so the
primary methods defined here in the package mask DESeq2.
}

\examples{
data(deseq)

## Get genes from DESeqDataSet.
dds <- as(deseq, "DESeqDataSet")
genes <- head(rownames(dds))
print(genes)

## DESeqAnalysis ====
plotMA(deseq, results = 1L)

## Customize the colors.
plotMA(
    object = deseq,
    results = 1L,
    pointColor = c(
        downregulated = "red",
        nonsignificant = "black",
        upregulated = "green"
    )
)

## Directional support (up or down).
plotMA(deseq, results = 1L, direction = "up", ntop = 5L)
plotMA(deseq, results = 1L, direction = "down", ntop = 5L)

## Label genes manually.
## Note that either gene IDs or names (symbols) are supported.
plotMA(deseq, results = 1L, genes = genes)
}
\seealso{
\code{\link[DESeq2:plotMA]{DESeq2::plotMA()}}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
