% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPCA-methods.R
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,DESeqAnalysis-method}
\title{Principal component analysis plot}
\usage{
plotPCA(object, ...)

\S4method{plotPCA}{DESeqAnalysis}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Principal component analysis plot
}
\details{
Passes to \code{SummarizedExperiment} defined in AcidPlots package.
}
\section{Functions}{
\itemize{
\item \code{plotPCA(DESeqAnalysis)}: Extracts \code{DESeqTransform}, converts to
\code{RangedSummarizedExperiment}, and passes to method defined in AcidPlots.

}}
\note{
Updated 2022-03-08.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotPCA(deseq)
}
