% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPca-methods.R
\name{plotPca}
\alias{plotPca}
\alias{plotPca,DESeqAnalysis-method}
\title{Principal component analysis plot}
\usage{
plotPca(object, ...)

\S4method{plotPca}{DESeqAnalysis}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Principal component analysis plot
}
\details{
Passes to \code{SummarizedExperiment} defined in AcidPlots package.
}
\section{Functions}{
\itemize{
\item \code{plotPca(DESeqAnalysis)}: Extracts \code{DESeqTransform}, converts to
\code{RangedSummarizedExperiment}, and passes to method defined in AcidPlots.

}}
\note{
Updated 2022-09-27.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotPca(deseq)
}
\seealso{
\itemize{
\item \code{DESeq2::plotPCA}.
}
}
