% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPCA-methods.R
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,DESeqAnalysis-method}
\title{Principal component analysis plot}
\usage{
plotPCA(object, ...)

\S4method{plotPCA}{DESeqAnalysis}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Principal component analysis plot
}
\details{
Passes to \code{SummarizedExperiment} defined in AcidPlots package.
}
\section{Functions}{
\itemize{
\item \code{plotPCA,DESeqAnalysis-method}: Extracts \code{DESeqTransform}, converts to
\code{RangedSummarizedExperiment}, and passes to method defined in AcidPlots.
}}

\note{
Updated 2022-01-21.
}
\section{Principal component analysis}{



PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.

}

\section{\code{SingleCellExperiment}}{



The \code{SingleCellExperiment} method that visualizes dimension reduction data
slotted in \code{reducedDims()} is defined in pointillism package.

}

\examples{
data(deseq)

## DESeqAnalysis ====
plotPCA(deseq)
}
\references{
Jolliffe, et al., 2002.
}
\seealso{
\itemize{
\item \code{DESeq2::plotPCA()}.
}
}
