% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotVolcano-methods.R
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqAnalysis-method}
\alias{plotVolcano,DESeqResults-method}
\title{Volcano plot}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqAnalysis}(
  object,
  i,
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  genes = NULL,
  ntop = 0L,
  ...
)

\S4method{plotVolcano}{DESeqResults}(
  object,
  direction = c("both", "up", "down"),
  alphaThreshold = NULL,
  baseMeanThreshold = NULL,
  lfcThreshold = NULL,
  genes = NULL,
  ntop = 0L,
  pointColor = c(downregulated = AcidPlots::lightPalette[["purple"]], upregulated =
    AcidPlots::lightPalette[["orange"]], nonsignificant =
    AcidPlots::lightPalette[["gray"]]),
  pointSize = 2L,
  pointAlpha = 0.8,
  limits = list(x = NULL, y = c(1e-10, 1L)),
  labels = list(title = NULL, subtitle = NULL),
  histograms = FALSE
)
}
\arguments{
\item{object}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{ntop}{\code{integer(1)}.
Number of top features to label.}

\item{...}{Additional arguments.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{limits}{\code{list(2)}.
Named list containing \code{"x"} and \code{"y"} that define the lower and upper
limits for each axis. Set automatically by default when left \code{NULL}.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{histograms}{\code{logical(1)}.
Show LFC and P value histograms.}
}
\value{
\code{ggplot}.
}
\description{
Volcano plot
}
\section{Functions}{
\itemize{
\item \code{plotVolcano,DESeqAnalysis-method}: Passes to \code{DESeqResults} method, with \code{gene2symbol}
argument automatically defined.
}}

\note{
Updated 2021-08-09.
}
\examples{
data(deseq)

## Get genes from DESeqDataSet.
dds <- as(deseq, "DESeqDataSet")
genes <- head(rownames(dds))
print(genes)

## DESeqAnalysis ====
plotVolcano(deseq, i = 1L)

## Customize the colors.
plotVolcano(
    object = deseq,
    i = 1L,
    pointColor = c(
        downregulated = "red",
        nonsignificant = "black",
        upregulated = "green"
    )
)

## Directional support (up or down).
plotVolcano(
    object = deseq,
    i = 1L,
    direction = "up",
    ntop = 5L
)
plotVolcano(
    object = deseq,
    i = 1L,
    direction = "down",
    ntop = 5L
)

## Label genes manually.
## Note that either gene IDs or names (symbols) are supported.
plotVolcano(deseq, i = 1L, genes = genes)
}
\seealso{
\code{CHBUtils::volcano_density_plot()}.
}
\author{
Michael Steinbaugh, John Hutchinson, Lorena Pantano
}
