% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano-methods.R
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\alias{plotVolcano,DESeqAnalysis-method}
\title{Volcano plot}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqResults}(
  object,
  ylim = 1e-10,
  genes = NULL,
  gene2symbol = NULL,
  ntop = 0L,
  direction = c("both", "up", "down"),
  pointColor = c(downregulated = acidplots::lightPalette[["purple"]], upregulated =
    acidplots::lightPalette[["orange"]], nonsignificant =
    acidplots::lightPalette[["gray"]]),
  pointSize = 2L,
  pointAlpha = 0.8,
  histograms = FALSE,
  return = c("ggplot", "DataFrame")
)

\S4method{plotVolcano}{DESeqAnalysis}(object, i, lfcShrink = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{ylim}{\code{numeric(1)}.
Upper boundary limit for y-axis. Helps preserve dynamic range for gene sets
containing highly significant P values (e.g. \code{1e-100}).}

\item{genes}{\code{character}.
Gene identifiers.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{ntop}{\code{integer(1)}.
Number of top features to label.}

\item{direction}{\code{character(1)}.
Plot \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{pointColor}{\code{character(1)}.
Default point color for the plot.}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{histograms}{\code{logical(1)}.
Show LFC and P value histograms.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Volcano plot
}
\section{Functions}{
\itemize{
\item \code{plotVolcano,DESeqAnalysis-method}: Passes to \code{DESeqResults} method, with \code{gene2symbol}
argument automatically defined.
}}

\note{
Updated 2019-11-19.
}
\examples{
data(deseq)

## Get genes from DESeqDataSet.
dds <- as(deseq, "DESeqDataSet")
genes <- head(rownames(dds))
print(genes)

## DESeqAnalysis ====
plotVolcano(deseq, i = 1L)

## Customize the colors.
plotVolcano(
    object = deseq,
    i = 1L,
    pointColor = c(
        downregulated = "red",
        nonsignificant = "black",
        upregulated = "green"
    )
)

## Directional support (up or down).
plotVolcano(
    object = deseq,
    i = 1L,
    direction = "up",
    ntop = 5L
)
plotVolcano(
    object = deseq,
    i = 1L,
    direction = "down",
    ntop = 5L
)

## Label genes manually.
## Note that either gene IDs or names (symbols) are supported.
plotVolcano(deseq, i = 1L, genes = genes)
}
\seealso{
\code{CHBUtils::volcano_density_plot()}.
}
\author{
Michael Steinbaugh, John Hutchinson, Lorena Pantano
}
