% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTximportFiles.R
\name{prepareTximportFiles}
\alias{prepareTximportFiles}
\title{Prepare quant files for tximport}
\usage{
prepareTximportFiles(
  files,
  makeNames = c("makeNames", "snakeCase", "camelCase"),
  exists = TRUE
)
}
\arguments{
\item{files}{\code{character}.
Quant file paths (e.g. "quant.sf" for salmon, "abundance.h5" for kallisto).
See \code{\link[tximport:tximport]{tximport()}} for details.}

\item{makeNames}{\code{character(1)}.
Syntactic name function to apply on sample names.
Uses \code{\link[base:match.arg]{match.arg()}} internally.
See basejump toolkit for details.}

\item{exists}{\code{logical(1)}.
Check if requested input exists on disk.
Runs \code{\link[basejump:reexports]{realpath()}} internally.}
}
\value{
\code{character}.
Return quant file paths, with valid sample names automatically applied.
}
\description{
Prepare quant files for tximport
}
\details{
Runs the following internal comments:
\enumerate{
\item Extract sample directory name from quant file using
\code{\link[base:basename]{dirname()}} and \code{\link[base:basename]{basename()}}.
\item Autopad zeros, if necessary, via
\code{\link[basejump:reexports]{autopadZeros()}}.
\item Sanitizes names with \code{\link[syntactic:snakeCase]{snakeCase()}}.
\item Sorts files alphabetically.
}
}
\note{
Updated 2022-03-23.
}
\examples{
files <- c(
    file.path("salmon", "1-sample-A", "quant.sf"),
    file.path("salmon", "2-sample-B", "quant.sf")
)
print(files)
files <- prepareTximportFiles(files, makeNames = "snakeCase", exists = FALSE)
print(files)
}
