% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSummary-methods.R,
%   R/contrastName-methods.R, R/export-methods.R, R/markdown-methods.R,
%   R/package.R, R/plotDEGHeatmap-methods.R, R/plotDEGPCA-methods.R,
%   R/plotGene-methods.R, R/plotMA-methods.R, R/plotVolcano-methods.R,
%   R/show-methods.R, R/topTables-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Gene2Symbol}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{alphaSummary}}, \code{\link[basejump]{contrastName}}, \code{\link[basejump]{export}}, \code{\link[basejump]{markdown}}, \code{\link[basejump]{Gene2Symbol}}, \code{\link[basejump]{plotDEGHeatmap}}, \code{\link[basejump]{plotDEGPCA}}, \code{\link[basejump]{plotGene}}, \code{\link[basejump]{plotVolcano}}, \code{\link[basejump]{topTables}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{plotMA}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

