% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,DESeqDataSet-method}
\alias{results,DESeqAnalysis-method}
\title{Results}
\usage{
\S4method{results}{DESeqDataSet}(object, ...)

\S4method{results}{DESeqAnalysis}(object, results, lfcShrink = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}
}
\description{
Results
}
