% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,DESeqDataSet-method}
\alias{results,DESeqAnalysis-method}
\title{Results}
\usage{
results(object, ...)

\S4method{results}{DESeqDataSet}(object, ...)

\S4method{results}{DESeqAnalysis}(object, i, lfcShrink = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}
}
\description{
Results
}
\note{
Updated 2019-11-08.
}
\examples{
data(deseq)
x <- results(deseq, i = 1L)
class(x)
}
