% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsDiff-methods.R
\name{resultsDiff}
\alias{resultsDiff}
\alias{resultsDiff,DESeqResults,DESeqResults-method}
\alias{resultsDiff,DESeqAnalysis,missingOrNULL-method}
\title{Relative difference of results}
\usage{
resultsDiff(x, y, ...)

\S4method{resultsDiff}{DESeqResults,DESeqResults}(
  x,
  y,
  col = c("log2FoldChange", "stat"),
  deg = c("no", "intersect", "union"),
  alphaThreshold = NULL,
  lfcThreshold = NULL,
  baseMeanThreshold = NULL,
  direction = c("both", "up", "down")
)

\S4method{resultsDiff}{DESeqAnalysis,missingOrNULL}(x, y = NULL, i, j, ...)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{col}{\code{character(1)}.
Column to use for difference calculation (subtraction).
Note that \code{x}/\code{i} represents \code{A} and \code{y}/\code{j} \code{B} in \code{A - B} calculation.}

\item{deg}{\code{character(1)}.
\itemize{
\item \code{"no"}: Return all genes, corresponding to original row names order.
\item \code{"intersect"}: Return only genes called as DE in both contrasts.
\item \code{"union"}: Return genes called as DE in either contrast.
}}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{direction}{\code{character(1)}.
Include \code{"both"}, \code{"up"}, or \code{"down"} directions.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{...}{Additional arguments.}
}
\value{
Named \code{numeric}.
Names correspond to the features (i.e. genes) analyzed.
}
\description{
Relative difference of results
}
\section{Functions}{
\itemize{
\item \code{resultsDiff,DESeqAnalysis,missingOrNULL-method}: Passes arguments to \code{DESeqResults} method.
}}

\note{
Updated 2020-08-05.
}
\examples{
data(deseq)

## DESeqAnalysis ====
diff <- resultsDiff(deseq, i = 1L, j = 2L)
head(diff)
}
