% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMatrix-methods.R
\docType{methods}
\name{resultsMatrix}
\alias{resultsMatrix}
\alias{resultsMatrix,DESeqAnalysis-method}
\alias{resultsMatrix,DESeqAnalysisList-method}
\title{DESeqResults matrix}
\usage{
resultsMatrix(object, ...)

\S4method{resultsMatrix}{DESeqAnalysis}(object,
  value = c("log2FoldChange", "stat", "padj"))

\S4method{resultsMatrix}{DESeqAnalysisList}(object,
  value = c("log2FoldChange", "stat", "padj"))
}
\arguments{
\item{object}{\code{DESeqAnalysis}.}

\item{value}{\code{character(1)}.
Value type to return. Corresponds to supported \code{DESeqResults} column:
\itemize{
\item \code{log2FoldChange}: log2 fold change. This will return \emph{shrunken} LFC
values if they are slotted in the \code{DESeqAnalysis} object.
\item \code{stat}: Wald test statistic.
\item \code{padj}: BH adjusted \emph{P} value.
}}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Generate an aggregate matrix of DESeqResults values.
}
\examples{
data(deseq)
x <- resultsMatrix(deseq)
head(x)
}
