% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsNames-methods.R
\name{resultsNames}
\alias{resultsNames}
\alias{resultsNames<-}
\alias{resultsNames,DESeqDataSet-method}
\alias{resultsNames,DESeqAnalysis-method}
\alias{resultsNames<-,DESeqAnalysis,character-method}
\title{Results names}
\usage{
resultsNames(object, ...)

resultsNames(object, ...) <- value

\S4method{resultsNames}{DESeqDataSet}(object)

\S4method{resultsNames}{DESeqAnalysis}(object)

\S4method{resultsNames}{DESeqAnalysis,character}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Results names
}
\note{
Updated 2019-09-10.
}
\examples{
data(deseq)

## DESeqAnalysis ====
resultsNames(deseq)
}
\seealso{
\itemize{
\item \code{DESeq2::resultsNames()}.
}
}
