% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsNames-methods.R
\docType{methods}
\name{resultsNames}
\alias{resultsNames}
\alias{resultsNames,DESeqDataSet-method}
\alias{resultsNames,DESeqAnalysis-method}
\title{Results names}
\usage{
resultsNames(object, ...)

\S4method{resultsNames}{DESeqDataSet}(object)

\S4method{resultsNames}{DESeqAnalysis}(object)
}
\description{
Results names
}
\seealso{
\code{\link[DESeq2:resultsNames]{DESeq2::resultsNames()}}.
}
