% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsNames-methods.R
\docType{methods}
\name{resultsNames}
\alias{resultsNames}
\alias{resultsNames,DESeqDataSet-method}
\alias{resultsNames,DESeqAnalysis-method}
\title{Results names}
\usage{
resultsNames(object, ...)

\S4method{resultsNames}{DESeqDataSet}(object)

\S4method{resultsNames}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}.
}
\description{
Results names
}
\seealso{
\itemize{
\item \code{DESeq2::resultsNames()}.
}
}
