% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsTables-methods.R
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\alias{resultsTables,DESeqAnalysis-method}
\title{Results tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(
  object,
  alphaThreshold = NULL,
  lfcThreshold = NULL,
  baseMeanThreshold = NULL,
  return = c("tbl_df", "DataFrameList")
)

\S4method{resultsTables}{DESeqAnalysis}(object, i, extra = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{alphaThreshold}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log (base 2) fold change ratio cutoff threshold.
If left \code{NULL}, will use the cutoff defined in the object.}

\item{baseMeanThreshold}{\code{numeric(1)} or \code{NULL}.
Base mean (i.e. average expression across all samples) threshold.
If left \code{NULL}, will use the cutoff defined in the object.
Applies in general to DESeq2 RNA-seq differential expression output.}

\item{return}{\code{character(1)}.
Type of data frame to return as a list.
Uses \link[base:match.arg]{match.arg()}.
\itemize{
\item \code{DataFrameList}: Returns \code{DataFrameList} with row names.
\item \code{tbl_df}: Returns \code{list} of \code{tbl_df} containing \code{"rowname"} column.
}}

\item{i}{\code{character(1)} or \code{integer(1)}.
\code{DESeqResults} contrast name or position in \code{results} slot.}

\item{extra}{\code{logical(1)}.
Include row data (i.e. gene metadata) and normalized counts from the
internal \code{DESeqDataSet}.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
Named list containing subsets of \code{DESeqResults}.
}
\description{
Results tables
}
\details{
Generate tables summarizing the differential expression, with subsets for
differentially expressed genes (DEGs). DEG tables (i.e. everything except the
\code{all} table), are arranged by adjusted \emph{P} value.
}
\note{
Updated 2020-09-21.

It is generally recommended to not apply post hoc log fold change
cutoffs. If a specific effect size is desired, instead run
\code{DESeq2::results()} using the \code{lfcThreshold} parameter. Refer to the DESeq2
documentation and vignette for details.
}
\section{Tables}{

\itemize{
\item \code{all}: All genes, including genes without an adjusted \emph{P} value. This table
is unmodified, and the rows have not been re-arranged or subset. It is
suitable for gene set enrichment analysis (GSEA).
\item \code{up}: Upregulated genes.
\item \code{down}: Downregulated genes.
\item \code{both}: Bidirectional DEGs (up- and down-regulated). This table can be
used for overrepresentation testing but should NOT be used for GSEA.
}
}

\examples{
data(deseq)

## DESeqAnalysis ====
x <- resultsTables(deseq, i = 1L)
names(x)
}
