% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsTables-methods.R
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\alias{resultsTables,DESeqAnalysis-method}
\title{Results tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(
  object,
  DESeqDataSet = NULL,
  alpha = NULL,
  lfcThreshold = NULL,
  return = c("tbl_df", "DataFrameList"),
  ...
)

\S4method{resultsTables}{DESeqAnalysis}(
  object,
  i,
  lfcShrink = TRUE,
  extra = TRUE,
  alpha = NULL,
  lfcThreshold = NULL,
  return = c("tbl_df", "DataFrameList")
)
}
\arguments{
\item{object}{Object.}

\item{DESeqDataSet}{\code{DESeqDataSet} or \code{NULL}.}

\item{alpha}{\code{numeric(1)} or \code{NULL}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{lfcThreshold}{\code{numeric(1)} or \code{NULL}.
Log fold change ratio (base 2) cutoff threshold. If left \code{NULL}, will use
the cutoff defined in the object.}

\item{return}{\code{character(1)}.
Type of data frame to return as a list.
Uses \link[base:match.arg]{match.arg()}.
\itemize{
\item \code{DataFrameList}: Returns \code{DataFrameList} with row names.
\item \code{tbl_df}: Returns \code{list} of \code{tbl_df} containing \code{"rowname"} column.
}}

\item{...}{Additional arguments.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}

\item{extra}{\code{logical(1)}.
Include row data and normalized counts from internal \code{DESeqDataSet}.}
}
\value{
\code{list}.
Named list containing subsets of \code{DESeqResults}.
}
\description{
Results tables
}
\details{
Generate tables summarizing the differential expression, with subsets for
differentially expressed genes (DEGs). DEG tables (i.e. everything except the
\code{all} table), are arranged by adjusted \emph{P} value.
}
\note{
Updated 2019-11-12.

Do not apply post hoc log fold change cutoffs.
}
\section{Tables}{

\itemize{
\item \code{all}: All genes, including genes without an adjusted \emph{P} value. This table
is unmodified, and the rows have not been re-arranged or subset. It is
suitable for gene set enrichment analysis (GSEA).
\item \code{up}: Upregulated genes.
\item \code{down}: Downregulated genes.
\item \code{both}: Bidirectional DEGs (up- and down-regulated). This table can be
used for overrepresentation testing but should NOT be used for GSEA.
}
}

\examples{
data(deseq)

## DESeqAnalysis ====
x <- resultsTables(deseq, i = 1L)
names(x)

## DESeqResults ====
## Use of DESeqAnalysis is encouraged instead of this approach.
res <- results(deseq, i = 1L)
dds <- as(deseq, "DESeqDataSet")
x <- resultsTables(object = res, DESeqDataSet = dds)
names(x)
}
