% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempusToTximport.R
\name{tempusToTximport}
\alias{tempusToTximport}
\title{Import Tempus transcript-level data into tximport}
\usage{
tempusToTximport(file)
}
\arguments{
\item{file}{\code{character(1)}.
Tempus transcript-level counts file.}
}
\value{
tximport \code{list}.
}
\description{
Import Tempus transcript-level data into tximport
}
\note{
Updated 2023-08-30.
}
\examples{
## > file <- "g_rna_abundance.csv"
## > txi <- tempusToTximport(file)
}
