% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables-methods.R
\docType{methods}
\name{topTables}
\alias{topTables}
\alias{topTables,DESeqResults-method}
\alias{topTables,DESeqResultsTables-method}
\alias{topTables,DESeqAnalysis-method}
\title{Top Tables of Differential Expression Results}
\usage{
\S4method{topTables}{DESeqResults}(object, n = 10L)

\S4method{topTables}{DESeqResultsTables}(object, n = 10L)

\S4method{topTables}{DESeqAnalysis}(object, results = 1L,
  lfcShrink = TRUE, n = 10L)
}
\arguments{
\item{object}{Object.}

\item{n}{\code{scalar integer}. Number of genes (per direction) to report.}
}
\value{
\code{kable}. Markdown tables.
}
\description{
Top Tables of Differential Expression Results
}
\examples{
data(deseq)
topTables(deseq, results = 1L, n = 5L)
}
