% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables-methods.R
\docType{methods}
\name{topTables}
\alias{topTables}
\alias{topTables,DESeqResults-method}
\alias{topTables,DESeqAnalysis-method}
\title{Top tables}
\usage{
topTables(object, ...)

\S4method{topTables}{DESeqResults}(object, n = 10L)

\S4method{topTables}{DESeqAnalysis}(object, results, n = 10L,
  lfcShrink = TRUE)
}
\arguments{
\item{object}{Object.}

\item{n}{\code{integer(1)}.
Number of genes (per direction) to report.}

\item{results}{\code{character(1)} or \code{integer(1)}.
Name or position of \code{DESeqResults}.}

\item{lfcShrink}{\code{logical(1)}.
Use shrunken log2 fold change (LFC) values.}

\item{...}{Additional arguments.}
}
\value{
\code{kable}.
Markdown tables.
}
\description{
Top tables of differentially expressed genes.
}
\examples{
data(deseq)

## DESeqAnalysis ====
topTables(deseq, results = 1L, n = 5L)
}
