% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transformType-methods.R
\docType{methods}
\name{transformType}
\alias{transformType}
\alias{transformType,DESeqTransform-method}
\alias{transformType,DESeqAnalysis-method}
\title{Variance-stabilizing transformation type}
\usage{
transformType(object, ...)

\S4method{transformType}{DESeqTransform}(object)

\S4method{transformType}{DESeqAnalysis}(object)
}
\description{
Variance-stabilizing transformation type
}
\seealso{
\code{\link[DESeq2:DESeqTransform]{DESeq2::DESeqTransform()}}.
}
