% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transformType-methods.R
\name{transformType}
\alias{transformType}
\alias{transformType,DESeqTransform-method}
\alias{transformType,DESeqAnalysis-method}
\title{Variance-stabilizing transformation type}
\usage{
transformType(object, ...)

\S4method{transformType}{DESeqTransform}(object)

\S4method{transformType}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Transformation method name.
}
\description{
Variance-stabilizing transformation type
}
\note{
Updated 2020-08-04.
}
\examples{
data(deseq)

## DESeqAnalysis ====
transformType(deseq)
}
\seealso{
\code{\link[DESeq2:DESeqTransform]{DESeq2::DESeqTransform()}}.
}
